Verify that the concentration and time are valid

If the concentrations or times are invalid, will provide an error. Reasons for being invalid are

time is not a number

conc is not a number

Any time value is NA

time is not monotonically increasing

conc and time are not the same length

check.conc.time(conc, time, monotonic.time = TRUE)

Arguments

conc

Measured concentrations

time

Time of the measurement of the concentrations

monotonic.time

Must the time be unique and monotonically increasing?

Value

None

Details

Some cases may generate warnings but allow the data to proceed.

A negative concentration is often but not always an error; it will generate a warning.