Get the columns that can be used in an interval specification

Get the columns that can be used in an interval specification

get.interval.cols()

Value

A list with named elements for each parameter. Each list element contains the parameter definition.

See also

Examples

get.interval.cols()

#> $start
#> $start$FUN
#> [1] NA
#> 
#> $start$values
#> function (x, ...)  .Primitive("as.double")
#> 
#> $start$desc
#> [1] "Starting time of the interval"
#> 
#> $start$formalsmap
#> list()
#> 
#> $start$depends
#> NULL
#> 
#> $start$datatype
#> [1] "interval"
#> 
#> 
#> $end
#> $end$FUN
#> [1] NA
#> 
#> $end$values
#> function (x, ...)  .Primitive("as.double")
#> 
#> $end$desc
#> [1] "Ending time of the interval (potentially infinity)"
#> 
#> $end$formalsmap
#> list()
#> 
#> $end$depends
#> NULL
#> 
#> $end$datatype
#> [1] "interval"
#> 
#> 
#> $auclast
#> $auclast$FUN
#> [1] "pk.calc.auc.last"
#> 
#> $auclast$values
#> [1] FALSE  TRUE
#> 
#> $auclast$desc
#> [1] "The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification"
#> 
#> $auclast$formalsmap
#> list()
#> 
#> $auclast$depends
#> NULL
#> 
#> $auclast$datatype
#> [1] "interval"
#> 
#> 
#> $aucall
#> $aucall$FUN
#> [1] "pk.calc.auc.all"
#> 
#> $aucall$values
#> [1] FALSE  TRUE
#> 
#> $aucall$desc
#> [1] "The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification"
#> 
#> $aucall$formalsmap
#> list()
#> 
#> $aucall$depends
#> NULL
#> 
#> $aucall$datatype
#> [1] "interval"
#> 
#> 
#> $aumclast
#> $aumclast$FUN
#> [1] "pk.calc.aumc.last"
#> 
#> $aumclast$values
#> [1] FALSE  TRUE
#> 
#> $aumclast$desc
#> [1] "The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification"
#> 
#> $aumclast$formalsmap
#> list()
#> 
#> $aumclast$depends
#> NULL
#> 
#> $aumclast$datatype
#> [1] "interval"
#> 
#> 
#> $aumcall
#> $aumcall$FUN
#> [1] "pk.calc.aumc.all"
#> 
#> $aumcall$values
#> [1] FALSE  TRUE
#> 
#> $aumcall$desc
#> [1] "The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification"
#> 
#> $aumcall$formalsmap
#> list()
#> 
#> $aumcall$depends
#> NULL
#> 
#> $aumcall$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.last
#> $aucint.last$FUN
#> [1] "pk.calc.aucint.last"
#> 
#> $aucint.last$values
#> [1] FALSE  TRUE
#> 
#> $aucint.last$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)"
#> 
#> $aucint.last$formalsmap
#> $aucint.last$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.last$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.last$formalsmap$time.dose
#> NULL
#> 
#> 
#> $aucint.last$depends
#> NULL
#> 
#> $aucint.last$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.last.dose
#> $aucint.last.dose$FUN
#> [1] "pk.calc.aucint.last"
#> 
#> $aucint.last.dose$values
#> [1] FALSE  TRUE
#> 
#> $aucint.last.dose$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)"
#> 
#> $aucint.last.dose$formalsmap
#> $aucint.last.dose$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.last.dose$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.last.dose$formalsmap$time.dose
#> [1] "time.dose.group"
#> 
#> 
#> $aucint.last.dose$depends
#> NULL
#> 
#> $aucint.last.dose$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.all
#> $aucint.all$FUN
#> [1] "pk.calc.aucint.all"
#> 
#> $aucint.all$values
#> [1] FALSE  TRUE
#> 
#> $aucint.all$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)"
#> 
#> $aucint.all$formalsmap
#> $aucint.all$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.all$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.all$formalsmap$time.dose
#> NULL
#> 
#> 
#> $aucint.all$depends
#> NULL
#> 
#> $aucint.all$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.all.dose
#> $aucint.all.dose$FUN
#> [1] "pk.calc.aucint.all"
#> 
#> $aucint.all.dose$values
#> [1] FALSE  TRUE
#> 
#> $aucint.all.dose$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)"
#> 
#> $aucint.all.dose$formalsmap
#> $aucint.all.dose$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.all.dose$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.all.dose$formalsmap$time.dose
#> [1] "time.dose.group"
#> 
#> 
#> $aucint.all.dose$depends
#> NULL
#> 
#> $aucint.all.dose$datatype
#> [1] "interval"
#> 
#> 
#> $auclast.dn
#> $auclast.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $auclast.dn$values
#> [1] FALSE  TRUE
#> 
#> $auclast.dn$desc
#> [1] "Dose normalized auclast"
#> 
#> $auclast.dn$formalsmap
#> $auclast.dn$formalsmap$parameter
#> [1] "auclast"
#> 
#> 
#> $auclast.dn$depends
#> [1] "auclast"
#> 
#> $auclast.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aucall.dn
#> $aucall.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aucall.dn$values
#> [1] FALSE  TRUE
#> 
#> $aucall.dn$desc
#> [1] "Dose normalized aucall"
#> 
#> $aucall.dn$formalsmap
#> $aucall.dn$formalsmap$parameter
#> [1] "aucall"
#> 
#> 
#> $aucall.dn$depends
#> [1] "aucall"
#> 
#> $aucall.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aumclast.dn
#> $aumclast.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aumclast.dn$values
#> [1] FALSE  TRUE
#> 
#> $aumclast.dn$desc
#> [1] "Dose normalized aumclast"
#> 
#> $aumclast.dn$formalsmap
#> $aumclast.dn$formalsmap$parameter
#> [1] "aumclast"
#> 
#> 
#> $aumclast.dn$depends
#> [1] "aumclast"
#> 
#> $aumclast.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aumcall.dn
#> $aumcall.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aumcall.dn$values
#> [1] FALSE  TRUE
#> 
#> $aumcall.dn$desc
#> [1] "Dose normalized aumcall"
#> 
#> $aumcall.dn$formalsmap
#> $aumcall.dn$formalsmap$parameter
#> [1] "aumcall"
#> 
#> 
#> $aumcall.dn$depends
#> [1] "aumcall"
#> 
#> $aumcall.dn$datatype
#> [1] "interval"
#> 
#> 
#> $cmax
#> $cmax$FUN
#> [1] "pk.calc.cmax"
#> 
#> $cmax$values
#> [1] FALSE  TRUE
#> 
#> $cmax$desc
#> [1] "Maximum observed concentration"
#> 
#> $cmax$formalsmap
#> list()
#> 
#> $cmax$depends
#> NULL
#> 
#> $cmax$datatype
#> [1] "interval"
#> 
#> 
#> $cmin
#> $cmin$FUN
#> [1] "pk.calc.cmin"
#> 
#> $cmin$values
#> [1] FALSE  TRUE
#> 
#> $cmin$desc
#> [1] "Minimum observed concentration"
#> 
#> $cmin$formalsmap
#> list()
#> 
#> $cmin$depends
#> NULL
#> 
#> $cmin$datatype
#> [1] "interval"
#> 
#> 
#> $tmax
#> $tmax$FUN
#> [1] "pk.calc.tmax"
#> 
#> $tmax$values
#> [1] FALSE  TRUE
#> 
#> $tmax$desc
#> [1] "Time of the maximum observed concentration"
#> 
#> $tmax$formalsmap
#> list()
#> 
#> $tmax$depends
#> NULL
#> 
#> $tmax$datatype
#> [1] "interval"
#> 
#> 
#> $tlast
#> $tlast$FUN
#> [1] "pk.calc.tlast"
#> 
#> $tlast$values
#> [1] FALSE  TRUE
#> 
#> $tlast$desc
#> [1] "Time of the last concentration observed above the limit of quantification"
#> 
#> $tlast$formalsmap
#> list()
#> 
#> $tlast$depends
#> NULL
#> 
#> $tlast$datatype
#> [1] "interval"
#> 
#> 
#> $tfirst
#> $tfirst$FUN
#> [1] "pk.calc.tfirst"
#> 
#> $tfirst$values
#> [1] FALSE  TRUE
#> 
#> $tfirst$desc
#> [1] "Time of the first concentration above the limit of quantification"
#> 
#> $tfirst$formalsmap
#> list()
#> 
#> $tfirst$depends
#> NULL
#> 
#> $tfirst$datatype
#> [1] "interval"
#> 
#> 
#> $clast.obs
#> $clast.obs$FUN
#> [1] "pk.calc.clast.obs"
#> 
#> $clast.obs$values
#> [1] FALSE  TRUE
#> 
#> $clast.obs$desc
#> [1] "The last concentration observed above the limit of quantification"
#> 
#> $clast.obs$formalsmap
#> list()
#> 
#> $clast.obs$depends
#> NULL
#> 
#> $clast.obs$datatype
#> [1] "interval"
#> 
#> 
#> $cl.last
#> $cl.last$FUN
#> [1] "pk.calc.cl"
#> 
#> $cl.last$values
#> [1] FALSE  TRUE
#> 
#> $cl.last$desc
#> [1] "Clearance or observed oral clearance calculated to Clast"
#> 
#> $cl.last$formalsmap
#> $cl.last$formalsmap$auc
#> [1] "auclast"
#> 
#> 
#> $cl.last$depends
#> [1] "auclast"
#> 
#> $cl.last$datatype
#> [1] "interval"
#> 
#> 
#> $cl.all
#> $cl.all$FUN
#> [1] "pk.calc.cl"
#> 
#> $cl.all$values
#> [1] FALSE  TRUE
#> 
#> $cl.all$desc
#> [1] "Clearance or observed oral clearance calculated with AUCall"
#> 
#> $cl.all$formalsmap
#> $cl.all$formalsmap$auc
#> [1] "aucall"
#> 
#> 
#> $cl.all$depends
#> [1] "aucall"
#> 
#> $cl.all$datatype
#> [1] "interval"
#> 
#> 
#> $f
#> $f$FUN
#> [1] "pk.calc.f"
#> 
#> $f$values
#> [1] FALSE  TRUE
#> 
#> $f$desc
#> [1] "Bioavailability or relative bioavailability"
#> 
#> $f$formalsmap
#> list()
#> 
#> $f$depends
#> NULL
#> 
#> $f$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.last
#> $mrt.last$FUN
#> [1] "pk.calc.mrt"
#> 
#> $mrt.last$values
#> [1] FALSE  TRUE
#> 
#> $mrt.last$desc
#> [1] "The mean residence time to the last observed concentration above the LOQ"
#> 
#> $mrt.last$formalsmap
#> $mrt.last$formalsmap$auc
#> [1] "auclast"
#> 
#> $mrt.last$formalsmap$aumc
#> [1] "aumclast"
#> 
#> 
#> $mrt.last$depends
#> $mrt.last$depends[[1]]
#> [1] "auclast"
#> 
#> $mrt.last$depends[[2]]
#> [1] "aumclast"
#> 
#> 
#> $mrt.last$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.iv.last
#> $mrt.iv.last$FUN
#> [1] "pk.calc.mrt.iv"
#> 
#> $mrt.iv.last$values
#> [1] FALSE  TRUE
#> 
#> $mrt.iv.last$desc
#> [1] "The mean residence time to the last observed concentration above the LOQ correcting for dosing duration"
#> 
#> $mrt.iv.last$formalsmap
#> $mrt.iv.last$formalsmap$auc
#> [1] "auclast"
#> 
#> $mrt.iv.last$formalsmap$aumc
#> [1] "aumclast"
#> 
#> 
#> $mrt.iv.last$depends
#> $mrt.iv.last$depends[[1]]
#> [1] "auclast"
#> 
#> $mrt.iv.last$depends[[2]]
#> [1] "aumclast"
#> 
#> 
#> $mrt.iv.last$datatype
#> [1] "interval"
#> 
#> 
#> $vss.last
#> $vss.last$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.last$values
#> [1] FALSE  TRUE
#> 
#> $vss.last$desc
#> [1] "The steady-state volume of distribution calculating through Tlast"
#> 
#> $vss.last$formalsmap
#> $vss.last$formalsmap$cl
#> [1] "cl.last"
#> 
#> $vss.last$formalsmap$mrt
#> [1] "mrt.last"
#> 
#> 
#> $vss.last$depends
#> [1] "cl.last"  "mrt.last"
#> 
#> $vss.last$datatype
#> [1] "interval"
#> 
#> 
#> $vss.iv.last
#> $vss.iv.last$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.iv.last$values
#> [1] FALSE  TRUE
#> 
#> $vss.iv.last$desc
#> [1] "The steady-state volume of distribution with intravenous infusion calculating through Tlast"
#> 
#> $vss.iv.last$formalsmap
#> $vss.iv.last$formalsmap$cl
#> [1] "cl.last"
#> 
#> $vss.iv.last$formalsmap$mrt
#> [1] "mrt.iv.last"
#> 
#> 
#> $vss.iv.last$depends
#> [1] "cl.last"     "mrt.iv.last"
#> 
#> $vss.iv.last$datatype
#> [1] "interval"
#> 
#> 
#> $cav
#> $cav$FUN
#> [1] "pk.calc.cav"
#> 
#> $cav$values
#> [1] FALSE  TRUE
#> 
#> $cav$desc
#> [1] "The average concentration during an interval"
#> 
#> $cav$formalsmap
#> list()
#> 
#> $cav$depends
#> [1] "auclast"
#> 
#> $cav$datatype
#> [1] "interval"
#> 
#> 
#> $ctrough
#> $ctrough$FUN
#> [1] "pk.calc.ctrough"
#> 
#> $ctrough$values
#> [1] FALSE  TRUE
#> 
#> $ctrough$desc
#> [1] "The trough (predose) concentration"
#> 
#> $ctrough$formalsmap
#> list()
#> 
#> $ctrough$depends
#> NULL
#> 
#> $ctrough$datatype
#> [1] "interval"
#> 
#> 
#> $ptr
#> $ptr$FUN
#> [1] "pk.calc.ptr"
#> 
#> $ptr$values
#> [1] FALSE  TRUE
#> 
#> $ptr$desc
#> [1] "Peak-to-Trough ratio (fraction)"
#> 
#> $ptr$formalsmap
#> list()
#> 
#> $ptr$depends
#> [1] "cmax"    "ctrough"
#> 
#> $ptr$datatype
#> [1] "interval"
#> 
#> 
#> $tlag
#> $tlag$FUN
#> [1] "pk.calc.tlag"
#> 
#> $tlag$values
#> [1] FALSE  TRUE
#> 
#> $tlag$desc
#> [1] "Lag time"
#> 
#> $tlag$formalsmap
#> list()
#> 
#> $tlag$depends
#> NULL
#> 
#> $tlag$datatype
#> [1] "interval"
#> 
#> 
#> $deg.fluc
#> $deg.fluc$FUN
#> [1] "pk.calc.deg.fluc"
#> 
#> $deg.fluc$values
#> [1] FALSE  TRUE
#> 
#> $deg.fluc$desc
#> [1] "Degree of fluctuation"
#> 
#> $deg.fluc$formalsmap
#> list()
#> 
#> $deg.fluc$depends
#> [1] "cmax" "cmin" "cav" 
#> 
#> $deg.fluc$datatype
#> [1] "interval"
#> 
#> 
#> $swing
#> $swing$FUN
#> [1] "pk.calc.swing"
#> 
#> $swing$values
#> [1] FALSE  TRUE
#> 
#> $swing$desc
#> [1] "Swing relative to Cmin"
#> 
#> $swing$formalsmap
#> list()
#> 
#> $swing$depends
#> [1] "cmax" "cmin"
#> 
#> $swing$datatype
#> [1] "interval"
#> 
#> 
#> $ceoi
#> $ceoi$FUN
#> [1] "pk.calc.ceoi"
#> 
#> $ceoi$values
#> [1] FALSE  TRUE
#> 
#> $ceoi$desc
#> [1] "Concentration at the end of infusion"
#> 
#> $ceoi$formalsmap
#> list()
#> 
#> $ceoi$depends
#> NULL
#> 
#> $ceoi$datatype
#> [1] "interval"
#> 
#> 
#> $ae
#> $ae$FUN
#> [1] "pk.calc.ae"
#> 
#> $ae$values
#> [1] FALSE  TRUE
#> 
#> $ae$desc
#> [1] "The amount excreted (typically into urine or feces)"
#> 
#> $ae$formalsmap
#> list()
#> 
#> $ae$depends
#> NULL
#> 
#> $ae$datatype
#> [1] "interval"
#> 
#> 
#> $clr.last
#> $clr.last$FUN
#> [1] "pk.calc.clr"
#> 
#> $clr.last$values
#> [1] FALSE  TRUE
#> 
#> $clr.last$desc
#> [1] "The renal clearance calculated using AUClast"
#> 
#> $clr.last$formalsmap
#> $clr.last$formalsmap$auc
#> [1] "auclast"
#> 
#> 
#> $clr.last$depends
#> NULL
#> 
#> $clr.last$datatype
#> [1] "interval"
#> 
#> 
#> $clr.obs
#> $clr.obs$FUN
#> [1] "pk.calc.clr"
#> 
#> $clr.obs$values
#> [1] FALSE  TRUE
#> 
#> $clr.obs$desc
#> [1] "The renal clearance calculated using AUCinf,obs"
#> 
#> $clr.obs$formalsmap
#> $clr.obs$formalsmap$auc
#> [1] "aucinf.obs"
#> 
#> 
#> $clr.obs$depends
#> NULL
#> 
#> $clr.obs$datatype
#> [1] "interval"
#> 
#> 
#> $clr.pred
#> $clr.pred$FUN
#> [1] "pk.calc.clr"
#> 
#> $clr.pred$values
#> [1] FALSE  TRUE
#> 
#> $clr.pred$desc
#> [1] "The renal clearance calculated using AUCinf,pred"
#> 
#> $clr.pred$formalsmap
#> $clr.pred$formalsmap$auc
#> [1] "aucinf.pred"
#> 
#> 
#> $clr.pred$depends
#> NULL
#> 
#> $clr.pred$datatype
#> [1] "interval"
#> 
#> 
#> $fe
#> $fe$FUN
#> [1] "pk.calc.fe"
#> 
#> $fe$values
#> [1] FALSE  TRUE
#> 
#> $fe$desc
#> [1] "The fraction of the dose excreted"
#> 
#> $fe$formalsmap
#> list()
#> 
#> $fe$depends
#> NULL
#> 
#> $fe$datatype
#> [1] "interval"
#> 
#> 
#> $time_above
#> $time_above$FUN
#> [1] "pk.calc.time_above"
#> 
#> $time_above$values
#> [1] FALSE  TRUE
#> 
#> $time_above$desc
#> [1] "Time above a given concentration"
#> 
#> $time_above$formalsmap
#> list()
#> 
#> $time_above$depends
#> NULL
#> 
#> $time_above$datatype
#> [1] "interval"
#> 
#> 
#> $half.life
#> $half.life$FUN
#> [1] "pk.calc.half.life"
#> 
#> $half.life$values
#> [1] FALSE  TRUE
#> 
#> $half.life$desc
#> [1] "The (terminal) half-life"
#> 
#> $half.life$formalsmap
#> list()
#> 
#> $half.life$depends
#> [1] "tmax"  "tlast"
#> 
#> $half.life$datatype
#> [1] "interval"
#> 
#> 
#> $r.squared
#> $r.squared$FUN
#> [1] NA
#> 
#> $r.squared$values
#> [1] FALSE  TRUE
#> 
#> $r.squared$desc
#> [1] "The r^2 value of the half-life calculation"
#> 
#> $r.squared$formalsmap
#> list()
#> 
#> $r.squared$depends
#> [1] "half.life"
#> 
#> $r.squared$datatype
#> [1] "interval"
#> 
#> 
#> $adj.r.squared
#> $adj.r.squared$FUN
#> [1] NA
#> 
#> $adj.r.squared$values
#> [1] FALSE  TRUE
#> 
#> $adj.r.squared$desc
#> [1] "The adjusted r^2 value of the half-life calculation"
#> 
#> $adj.r.squared$formalsmap
#> list()
#> 
#> $adj.r.squared$depends
#> [1] "half.life"
#> 
#> $adj.r.squared$datatype
#> [1] "interval"
#> 
#> 
#> $lambda.z
#> $lambda.z$FUN
#> [1] NA
#> 
#> $lambda.z$values
#> [1] FALSE  TRUE
#> 
#> $lambda.z$desc
#> [1] "The elimination rate of the terminal half-life"
#> 
#> $lambda.z$formalsmap
#> list()
#> 
#> $lambda.z$depends
#> [1] "half.life"
#> 
#> $lambda.z$datatype
#> [1] "interval"
#> 
#> 
#> $lambda.z.time.first
#> $lambda.z.time.first$FUN
#> [1] NA
#> 
#> $lambda.z.time.first$values
#> [1] FALSE  TRUE
#> 
#> $lambda.z.time.first$desc
#> [1] "The first time point used for the calculation of half-life"
#> 
#> $lambda.z.time.first$formalsmap
#> list()
#> 
#> $lambda.z.time.first$depends
#> [1] "half.life"
#> 
#> $lambda.z.time.first$datatype
#> [1] "interval"
#> 
#> 
#> $lambda.z.n.points
#> $lambda.z.n.points$FUN
#> [1] NA
#> 
#> $lambda.z.n.points$values
#> [1] FALSE  TRUE
#> 
#> $lambda.z.n.points$desc
#> [1] "The number of points used for the calculation of half-life"
#> 
#> $lambda.z.n.points$formalsmap
#> list()
#> 
#> $lambda.z.n.points$depends
#> [1] "half.life"
#> 
#> $lambda.z.n.points$datatype
#> [1] "interval"
#> 
#> 
#> $clast.pred
#> $clast.pred$FUN
#> [1] NA
#> 
#> $clast.pred$values
#> [1] FALSE  TRUE
#> 
#> $clast.pred$desc
#> [1] "The concentration at Tlast as predicted by the half-life"
#> 
#> $clast.pred$formalsmap
#> list()
#> 
#> $clast.pred$depends
#> [1] "half.life"
#> 
#> $clast.pred$datatype
#> [1] "interval"
#> 
#> 
#> $span.ratio
#> $span.ratio$FUN
#> [1] NA
#> 
#> $span.ratio$values
#> [1] FALSE  TRUE
#> 
#> $span.ratio$desc
#> [1] "The ratio of the half-life to the duration used for half-life calculation"
#> 
#> $span.ratio$formalsmap
#> list()
#> 
#> $span.ratio$depends
#> [1] "half.life"
#> 
#> $span.ratio$datatype
#> [1] "interval"
#> 
#> 
#> $cmax.dn
#> $cmax.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $cmax.dn$values
#> [1] FALSE  TRUE
#> 
#> $cmax.dn$desc
#> [1] "Dose normalized cmax"
#> 
#> $cmax.dn$formalsmap
#> $cmax.dn$formalsmap$parameter
#> [1] "cmax"
#> 
#> 
#> $cmax.dn$depends
#> [1] "cmax"
#> 
#> $cmax.dn$datatype
#> [1] "interval"
#> 
#> 
#> $cmin.dn
#> $cmin.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $cmin.dn$values
#> [1] FALSE  TRUE
#> 
#> $cmin.dn$desc
#> [1] "Dose normalized cmin"
#> 
#> $cmin.dn$formalsmap
#> $cmin.dn$formalsmap$parameter
#> [1] "cmin"
#> 
#> 
#> $cmin.dn$depends
#> [1] "cmin"
#> 
#> $cmin.dn$datatype
#> [1] "interval"
#> 
#> 
#> $clast.obs.dn
#> $clast.obs.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $clast.obs.dn$values
#> [1] FALSE  TRUE
#> 
#> $clast.obs.dn$desc
#> [1] "Dose normalized clast.obs"
#> 
#> $clast.obs.dn$formalsmap
#> $clast.obs.dn$formalsmap$parameter
#> [1] "clast.obs"
#> 
#> 
#> $clast.obs.dn$depends
#> [1] "clast.obs"
#> 
#> $clast.obs.dn$datatype
#> [1] "interval"
#> 
#> 
#> $clast.pred.dn
#> $clast.pred.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $clast.pred.dn$values
#> [1] FALSE  TRUE
#> 
#> $clast.pred.dn$desc
#> [1] "Dose normalized clast.pred"
#> 
#> $clast.pred.dn$formalsmap
#> $clast.pred.dn$formalsmap$parameter
#> [1] "clast.pred"
#> 
#> 
#> $clast.pred.dn$depends
#> [1] "clast.pred"
#> 
#> $clast.pred.dn$datatype
#> [1] "interval"
#> 
#> 
#> $cav.dn
#> $cav.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $cav.dn$values
#> [1] FALSE  TRUE
#> 
#> $cav.dn$desc
#> [1] "Dose normalized cav"
#> 
#> $cav.dn$formalsmap
#> $cav.dn$formalsmap$parameter
#> [1] "cav"
#> 
#> 
#> $cav.dn$depends
#> [1] "cav"
#> 
#> $cav.dn$datatype
#> [1] "interval"
#> 
#> 
#> $ctrough.dn
#> $ctrough.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $ctrough.dn$values
#> [1] FALSE  TRUE
#> 
#> $ctrough.dn$desc
#> [1] "Dose normalized ctrough"
#> 
#> $ctrough.dn$formalsmap
#> $ctrough.dn$formalsmap$parameter
#> [1] "ctrough"
#> 
#> 
#> $ctrough.dn$depends
#> [1] "ctrough"
#> 
#> $ctrough.dn$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.last
#> $thalf.eff.last$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.last$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.last$desc
#> [1] "The effective half-life (as determined from the MRTlast)"
#> 
#> $thalf.eff.last$formalsmap
#> $thalf.eff.last$formalsmap$mrt
#> [1] "mrt.last"
#> 
#> 
#> $thalf.eff.last$depends
#> [1] "mrt.last"
#> 
#> $thalf.eff.last$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.iv.last
#> $thalf.eff.iv.last$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.iv.last$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.iv.last$desc
#> [1] "The effective half-life (as determined from the intravenous MRTlast)"
#> 
#> $thalf.eff.iv.last$formalsmap
#> $thalf.eff.iv.last$formalsmap$mrt
#> [1] "mrt.iv.last"
#> 
#> 
#> $thalf.eff.iv.last$depends
#> [1] "mrt.iv.last"
#> 
#> $thalf.eff.iv.last$datatype
#> [1] "interval"
#> 
#> 
#> $kel.last
#> $kel.last$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.last$values
#> [1] FALSE  TRUE
#> 
#> $kel.last$desc
#> [1] "Elimination rate (as calculated from the MRT using AUClast)"
#> 
#> $kel.last$formalsmap
#> $kel.last$formalsmap$mrt
#> [1] "mrt.last"
#> 
#> 
#> $kel.last$depends
#> [1] "mrt.last"
#> 
#> $kel.last$datatype
#> [1] "interval"
#> 
#> 
#> $kel.iv.last
#> $kel.iv.last$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.iv.last$values
#> [1] FALSE  TRUE
#> 
#> $kel.iv.last$desc
#> [1] "Elimination rate (as calculated from the intravenous MRTlast)"
#> 
#> $kel.iv.last$formalsmap
#> $kel.iv.last$formalsmap$mrt
#> [1] "mrt.iv.last"
#> 
#> 
#> $kel.iv.last$depends
#> [1] "mrt.iv.last"
#> 
#> $kel.iv.last$datatype
#> [1] "interval"
#> 
#> 
#> $aucinf.obs
#> $aucinf.obs$FUN
#> [1] "pk.calc.auc.inf.obs"
#> 
#> $aucinf.obs$values
#> [1] FALSE  TRUE
#> 
#> $aucinf.obs$desc
#> [1] "The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast"
#> 
#> $aucinf.obs$formalsmap
#> list()
#> 
#> $aucinf.obs$depends
#> [1] "lambda.z"  "clast.obs"
#> 
#> $aucinf.obs$datatype
#> [1] "interval"
#> 
#> 
#> $aucinf.pred
#> $aucinf.pred$FUN
#> [1] "pk.calc.auc.inf.pred"
#> 
#> $aucinf.pred$values
#> [1] FALSE  TRUE
#> 
#> $aucinf.pred$desc
#> [1] "The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast"
#> 
#> $aucinf.pred$formalsmap
#> list()
#> 
#> $aucinf.pred$depends
#> [1] "lambda.z"   "clast.pred"
#> 
#> $aucinf.pred$datatype
#> [1] "interval"
#> 
#> 
#> $aumcinf.obs
#> $aumcinf.obs$FUN
#> [1] "pk.calc.aumc.inf.obs"
#> 
#> $aumcinf.obs$values
#> [1] FALSE  TRUE
#> 
#> $aumcinf.obs$desc
#> [1] "The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast"
#> 
#> $aumcinf.obs$formalsmap
#> list()
#> 
#> $aumcinf.obs$depends
#> [1] "lambda.z"  "clast.obs"
#> 
#> $aumcinf.obs$datatype
#> [1] "interval"
#> 
#> 
#> $aumcinf.pred
#> $aumcinf.pred$FUN
#> [1] "pk.calc.aumc.inf.pred"
#> 
#> $aumcinf.pred$values
#> [1] FALSE  TRUE
#> 
#> $aumcinf.pred$desc
#> [1] "The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast"
#> 
#> $aumcinf.pred$formalsmap
#> list()
#> 
#> $aumcinf.pred$depends
#> [1] "lambda.z"   "clast.pred"
#> 
#> $aumcinf.pred$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.inf.obs
#> $aucint.inf.obs$FUN
#> [1] "pk.calc.aucint.inf.obs"
#> 
#> $aucint.inf.obs$values
#> [1] FALSE  TRUE
#> 
#> $aucint.inf.obs$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)"
#> 
#> $aucint.inf.obs$formalsmap
#> $aucint.inf.obs$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.inf.obs$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.inf.obs$formalsmap$time.dose
#> NULL
#> 
#> 
#> $aucint.inf.obs$depends
#> [1] "lambda.z"  "clast.obs"
#> 
#> $aucint.inf.obs$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.inf.obs.dose
#> $aucint.inf.obs.dose$FUN
#> [1] "pk.calc.aucint.inf.obs"
#> 
#> $aucint.inf.obs.dose$values
#> [1] FALSE  TRUE
#> 
#> $aucint.inf.obs.dose$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)"
#> 
#> $aucint.inf.obs.dose$formalsmap
#> $aucint.inf.obs.dose$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.inf.obs.dose$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.inf.obs.dose$formalsmap$time.dose
#> [1] "time.dose.group"
#> 
#> 
#> $aucint.inf.obs.dose$depends
#> [1] "lambda.z"  "clast.obs"
#> 
#> $aucint.inf.obs.dose$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.inf.pred
#> $aucint.inf.pred$FUN
#> [1] "pk.calc.aucint.inf.pred"
#> 
#> $aucint.inf.pred$values
#> [1] FALSE  TRUE
#> 
#> $aucint.inf.pred$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)"
#> 
#> $aucint.inf.pred$formalsmap
#> $aucint.inf.pred$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.inf.pred$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.inf.pred$formalsmap$time.dose
#> NULL
#> 
#> 
#> $aucint.inf.pred$depends
#> [1] "lambda.z"   "clast.pred"
#> 
#> $aucint.inf.pred$datatype
#> [1] "interval"
#> 
#> 
#> $aucint.inf.pred.dose
#> $aucint.inf.pred.dose$FUN
#> [1] "pk.calc.aucint.inf.pred"
#> 
#> $aucint.inf.pred.dose$values
#> [1] FALSE  TRUE
#> 
#> $aucint.inf.pred.dose$desc
#> [1] "The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)"
#> 
#> $aucint.inf.pred.dose$formalsmap
#> $aucint.inf.pred.dose$formalsmap$conc
#> [1] "conc.group"
#> 
#> $aucint.inf.pred.dose$formalsmap$time
#> [1] "time.group"
#> 
#> $aucint.inf.pred.dose$formalsmap$time.dose
#> [1] "time.dose.group"
#> 
#> 
#> $aucint.inf.pred.dose$depends
#> [1] "lambda.z"   "clast.pred"
#> 
#> $aucint.inf.pred.dose$datatype
#> [1] "interval"
#> 
#> 
#> $aucinf.obs.dn
#> $aucinf.obs.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aucinf.obs.dn$values
#> [1] FALSE  TRUE
#> 
#> $aucinf.obs.dn$desc
#> [1] "Dose normalized aucinf.obs"
#> 
#> $aucinf.obs.dn$formalsmap
#> $aucinf.obs.dn$formalsmap$parameter
#> [1] "aucinf.obs"
#> 
#> 
#> $aucinf.obs.dn$depends
#> [1] "aucinf.obs"
#> 
#> $aucinf.obs.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aucinf.pred.dn
#> $aucinf.pred.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aucinf.pred.dn$values
#> [1] FALSE  TRUE
#> 
#> $aucinf.pred.dn$desc
#> [1] "Dose normalized aucinf.pred"
#> 
#> $aucinf.pred.dn$formalsmap
#> $aucinf.pred.dn$formalsmap$parameter
#> [1] "aucinf.pred"
#> 
#> 
#> $aucinf.pred.dn$depends
#> [1] "aucinf.pred"
#> 
#> $aucinf.pred.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aumcinf.obs.dn
#> $aumcinf.obs.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aumcinf.obs.dn$values
#> [1] FALSE  TRUE
#> 
#> $aumcinf.obs.dn$desc
#> [1] "Dose normalized aumcinf.obs"
#> 
#> $aumcinf.obs.dn$formalsmap
#> $aumcinf.obs.dn$formalsmap$parameter
#> [1] "aumcinf.obs"
#> 
#> 
#> $aumcinf.obs.dn$depends
#> [1] "aumcinf.obs"
#> 
#> $aumcinf.obs.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aumcinf.pred.dn
#> $aumcinf.pred.dn$FUN
#> [1] "pk.calc.dn"
#> 
#> $aumcinf.pred.dn$values
#> [1] FALSE  TRUE
#> 
#> $aumcinf.pred.dn$desc
#> [1] "Dose normalized aumcinf.pred"
#> 
#> $aumcinf.pred.dn$formalsmap
#> $aumcinf.pred.dn$formalsmap$parameter
#> [1] "aumcinf.pred"
#> 
#> 
#> $aumcinf.pred.dn$depends
#> [1] "aumcinf.pred"
#> 
#> $aumcinf.pred.dn$datatype
#> [1] "interval"
#> 
#> 
#> $aucpext.obs
#> $aucpext.obs$FUN
#> [1] "pk.calc.aucpext"
#> 
#> $aucpext.obs$values
#> [1] FALSE  TRUE
#> 
#> $aucpext.obs$desc
#> [1] "Percent of the AUCinf that is extrapolated after Tlast calculated from the observed Clast"
#> 
#> $aucpext.obs$formalsmap
#> $aucpext.obs$formalsmap$aucinf
#> [1] "aucinf.obs"
#> 
#> 
#> $aucpext.obs$depends
#> [1] "auclast"    "aucinf.obs"
#> 
#> $aucpext.obs$datatype
#> [1] "interval"
#> 
#> 
#> $aucpext.pred
#> $aucpext.pred$FUN
#> [1] "pk.calc.aucpext"
#> 
#> $aucpext.pred$values
#> [1] FALSE  TRUE
#> 
#> $aucpext.pred$desc
#> [1] "Percent of the AUCinf that is extrapolated after Tlast calculated from the predicted Clast"
#> 
#> $aucpext.pred$formalsmap
#> $aucpext.pred$formalsmap$aucinf
#> [1] "aucinf.pred"
#> 
#> 
#> $aucpext.pred$depends
#> [1] "auclast"     "aucinf.pred"
#> 
#> $aucpext.pred$datatype
#> [1] "interval"
#> 
#> 
#> $cl.obs
#> $cl.obs$FUN
#> [1] "pk.calc.cl"
#> 
#> $cl.obs$values
#> [1] FALSE  TRUE
#> 
#> $cl.obs$desc
#> [1] "Clearance or observed oral clearance calculated with observed Clast"
#> 
#> $cl.obs$formalsmap
#> $cl.obs$formalsmap$auc
#> [1] "aucinf.obs"
#> 
#> 
#> $cl.obs$depends
#> [1] "aucinf.obs"
#> 
#> $cl.obs$datatype
#> [1] "interval"
#> 
#> 
#> $cl.pred
#> $cl.pred$FUN
#> [1] "pk.calc.cl"
#> 
#> $cl.pred$values
#> [1] FALSE  TRUE
#> 
#> $cl.pred$desc
#> [1] "Clearance or observed oral clearance calculated with predicted Clast"
#> 
#> $cl.pred$formalsmap
#> $cl.pred$formalsmap$auc
#> [1] "aucinf.pred"
#> 
#> 
#> $cl.pred$depends
#> $cl.pred$depends[[1]]
#> [1] "aucinf.pred"
#> 
#> 
#> $cl.pred$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.obs
#> $mrt.obs$FUN
#> [1] "pk.calc.mrt"
#> 
#> $mrt.obs$values
#> [1] FALSE  TRUE
#> 
#> $mrt.obs$desc
#> [1] "The mean residence time to infinity using observed Clast"
#> 
#> $mrt.obs$formalsmap
#> $mrt.obs$formalsmap$auc
#> [1] "aucinf.obs"
#> 
#> $mrt.obs$formalsmap$aumc
#> [1] "aumcinf.obs"
#> 
#> 
#> $mrt.obs$depends
#> [1] "aucinf.obs"  "aumcinf.obs"
#> 
#> $mrt.obs$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.pred
#> $mrt.pred$FUN
#> [1] "pk.calc.mrt"
#> 
#> $mrt.pred$values
#> [1] FALSE  TRUE
#> 
#> $mrt.pred$desc
#> [1] "The mean residence time to infinity using predicted Clast"
#> 
#> $mrt.pred$formalsmap
#> $mrt.pred$formalsmap$auc
#> [1] "aucinf.pred"
#> 
#> $mrt.pred$formalsmap$aumc
#> [1] "aumcinf.pred"
#> 
#> 
#> $mrt.pred$depends
#> [1] "aucinf.pred"  "aumcinf.pred"
#> 
#> $mrt.pred$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.iv.obs
#> $mrt.iv.obs$FUN
#> [1] "pk.calc.mrt.iv"
#> 
#> $mrt.iv.obs$values
#> [1] FALSE  TRUE
#> 
#> $mrt.iv.obs$desc
#> [1] "The mean residence time to infinity using observed Clast correcting for dosing duration"
#> 
#> $mrt.iv.obs$formalsmap
#> $mrt.iv.obs$formalsmap$auc
#> [1] "aucinf.obs"
#> 
#> $mrt.iv.obs$formalsmap$aumc
#> [1] "aumcinf.obs"
#> 
#> 
#> $mrt.iv.obs$depends
#> [1] "aucinf.obs"  "aumcinf.obs"
#> 
#> $mrt.iv.obs$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.iv.pred
#> $mrt.iv.pred$FUN
#> [1] "pk.calc.mrt.iv"
#> 
#> $mrt.iv.pred$values
#> [1] FALSE  TRUE
#> 
#> $mrt.iv.pred$desc
#> [1] "The mean residence time to infinity using predicted Clast correcting for dosing duration"
#> 
#> $mrt.iv.pred$formalsmap
#> $mrt.iv.pred$formalsmap$auc
#> [1] "aucinf.pred"
#> 
#> $mrt.iv.pred$formalsmap$aumc
#> [1] "aumcinf.pred"
#> 
#> 
#> $mrt.iv.pred$depends
#> [1] "aucinf.pred"  "aumcinf.pred"
#> 
#> $mrt.iv.pred$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.md.obs
#> $mrt.md.obs$FUN
#> [1] "pk.calc.mrt.md"
#> 
#> $mrt.md.obs$values
#> [1] FALSE  TRUE
#> 
#> $mrt.md.obs$desc
#> [1] "The mean residence time with multiple dosing and nonlinear kinetics using observed Clast"
#> 
#> $mrt.md.obs$formalsmap
#> $mrt.md.obs$formalsmap$auctau
#> [1] "auclast"
#> 
#> $mrt.md.obs$formalsmap$aumctau
#> [1] "aumclast"
#> 
#> $mrt.md.obs$formalsmap$aucinf
#> [1] "aucinf.obs"
#> 
#> 
#> $mrt.md.obs$depends
#> [1] "auclast"    "aumclast"   "aucinf.obs"
#> 
#> $mrt.md.obs$datatype
#> [1] "interval"
#> 
#> 
#> $mrt.md.pred
#> $mrt.md.pred$FUN
#> [1] "pk.calc.mrt.md"
#> 
#> $mrt.md.pred$values
#> [1] FALSE  TRUE
#> 
#> $mrt.md.pred$desc
#> [1] "The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast"
#> 
#> $mrt.md.pred$formalsmap
#> $mrt.md.pred$formalsmap$auctau
#> [1] "auclast"
#> 
#> $mrt.md.pred$formalsmap$aumctau
#> [1] "aumclast"
#> 
#> $mrt.md.pred$formalsmap$aucinf
#> [1] "aucinf.pred"
#> 
#> 
#> $mrt.md.pred$depends
#> [1] "auclast"     "aumclast"    "aucinf.pred"
#> 
#> $mrt.md.pred$datatype
#> [1] "interval"
#> 
#> 
#> $vz.obs
#> $vz.obs$FUN
#> [1] "pk.calc.vz"
#> 
#> $vz.obs$values
#> [1] FALSE  TRUE
#> 
#> $vz.obs$desc
#> [1] "The terminal volume of distribution using observed Clast"
#> 
#> $vz.obs$formalsmap
#> $vz.obs$formalsmap$cl
#> [1] "cl.obs"
#> 
#> 
#> $vz.obs$depends
#> [1] "cl.obs"   "lambda.z"
#> 
#> $vz.obs$datatype
#> [1] "interval"
#> 
#> 
#> $vz.pred
#> $vz.pred$FUN
#> [1] "pk.calc.vz"
#> 
#> $vz.pred$values
#> [1] FALSE  TRUE
#> 
#> $vz.pred$desc
#> [1] "The terminal volume of distribution using predicted Clast"
#> 
#> $vz.pred$formalsmap
#> $vz.pred$formalsmap$cl
#> [1] "cl.pred"
#> 
#> 
#> $vz.pred$depends
#> [1] "cl.pred"  "lambda.z"
#> 
#> $vz.pred$datatype
#> [1] "interval"
#> 
#> 
#> $vss.obs
#> $vss.obs$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.obs$values
#> [1] FALSE  TRUE
#> 
#> $vss.obs$desc
#> [1] "The steady-state volume of distribution using observed Clast"
#> 
#> $vss.obs$formalsmap
#> $vss.obs$formalsmap$cl
#> [1] "cl.obs"
#> 
#> $vss.obs$formalsmap$mrt
#> [1] "mrt.obs"
#> 
#> 
#> $vss.obs$depends
#> [1] "cl.obs"  "mrt.obs"
#> 
#> $vss.obs$datatype
#> [1] "interval"
#> 
#> 
#> $vss.pred
#> $vss.pred$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.pred$values
#> [1] FALSE  TRUE
#> 
#> $vss.pred$desc
#> [1] "The steady-state volume of distribution using predicted Clast"
#> 
#> $vss.pred$formalsmap
#> $vss.pred$formalsmap$cl
#> [1] "cl.pred"
#> 
#> $vss.pred$formalsmap$mrt
#> [1] "mrt.pred"
#> 
#> 
#> $vss.pred$depends
#> [1] "cl.pred"  "mrt.pred"
#> 
#> $vss.pred$datatype
#> [1] "interval"
#> 
#> 
#> $vss.iv.obs
#> $vss.iv.obs$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.iv.obs$values
#> [1] FALSE  TRUE
#> 
#> $vss.iv.obs$desc
#> [1] "The steady-state volume of distribution with intravenous infusion using observed Clast"
#> 
#> $vss.iv.obs$formalsmap
#> $vss.iv.obs$formalsmap$cl
#> [1] "cl.obs"
#> 
#> $vss.iv.obs$formalsmap$mrt
#> [1] "mrt.iv.obs"
#> 
#> 
#> $vss.iv.obs$depends
#> [1] "cl.obs"     "mrt.iv.obs"
#> 
#> $vss.iv.obs$datatype
#> [1] "interval"
#> 
#> 
#> $vss.iv.pred
#> $vss.iv.pred$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.iv.pred$values
#> [1] FALSE  TRUE
#> 
#> $vss.iv.pred$desc
#> [1] "The steady-state volume of distribution with intravenous infusion using predicted Clast"
#> 
#> $vss.iv.pred$formalsmap
#> $vss.iv.pred$formalsmap$cl
#> [1] "cl.pred"
#> 
#> $vss.iv.pred$formalsmap$mrt
#> [1] "mrt.iv.pred"
#> 
#> 
#> $vss.iv.pred$depends
#> [1] "cl.pred"     "mrt.iv.pred"
#> 
#> $vss.iv.pred$datatype
#> [1] "interval"
#> 
#> 
#> $vss.md.obs
#> $vss.md.obs$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.md.obs$values
#> [1] FALSE  TRUE
#> 
#> $vss.md.obs$desc
#> [1] "The steady-state volume of distribution for nonlinear multiple-dose data using observed Clast"
#> 
#> $vss.md.obs$formalsmap
#> $vss.md.obs$formalsmap$cl
#> [1] "cl.last"
#> 
#> $vss.md.obs$formalsmap$mrt
#> [1] "mrt.md.obs"
#> 
#> 
#> $vss.md.obs$depends
#> [1] "cl.last"    "mrt.md.obs"
#> 
#> $vss.md.obs$datatype
#> [1] "interval"
#> 
#> 
#> $vss.md.pred
#> $vss.md.pred$FUN
#> [1] "pk.calc.vss"
#> 
#> $vss.md.pred$values
#> [1] FALSE  TRUE
#> 
#> $vss.md.pred$desc
#> [1] "The steady-state volume of distribution for nonlinear multiple-dose data using predicted Clast"
#> 
#> $vss.md.pred$formalsmap
#> $vss.md.pred$formalsmap$cl
#> [1] "cl.last"
#> 
#> $vss.md.pred$formalsmap$mrt
#> [1] "mrt.md.pred"
#> 
#> 
#> $vss.md.pred$depends
#> [1] "cl.last"     "mrt.md.pred"
#> 
#> $vss.md.pred$datatype
#> [1] "interval"
#> 
#> 
#> $vd.obs
#> $vd.obs$FUN
#> [1] "pk.calc.vd"
#> 
#> $vd.obs$values
#> [1] FALSE  TRUE
#> 
#> $vd.obs$desc
#> [1] "Apparent observed volume of distribution calculated with observed Clast"
#> 
#> $vd.obs$formalsmap
#> $vd.obs$formalsmap$aucinf
#> [1] "aucinf.obs"
#> 
#> 
#> $vd.obs$depends
#> [1] "aucinf.obs" "lambda.z"  
#> 
#> $vd.obs$datatype
#> [1] "interval"
#> 
#> 
#> $vd.pred
#> $vd.pred$FUN
#> [1] "pk.calc.vd"
#> 
#> $vd.pred$values
#> [1] FALSE  TRUE
#> 
#> $vd.pred$desc
#> [1] "Apparent observed volume of distribution calculated with predicted Clast"
#> 
#> $vd.pred$formalsmap
#> $vd.pred$formalsmap$aucinf
#> [1] "aucinf.pred"
#> 
#> 
#> $vd.pred$depends
#> [1] "aucinf.pred" "lambda.z"   
#> 
#> $vd.pred$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.obs
#> $thalf.eff.obs$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.obs$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.obs$desc
#> [1] "The effective half-life (as determined from the MRTobs)"
#> 
#> $thalf.eff.obs$formalsmap
#> $thalf.eff.obs$formalsmap$mrt
#> [1] "mrt.obs"
#> 
#> 
#> $thalf.eff.obs$depends
#> [1] "mrt.obs"
#> 
#> $thalf.eff.obs$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.pred
#> $thalf.eff.pred$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.pred$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.pred$desc
#> [1] "The effective half-life (as determined from the MRTpred)"
#> 
#> $thalf.eff.pred$formalsmap
#> $thalf.eff.pred$formalsmap$mrt
#> [1] "mrt.pred"
#> 
#> 
#> $thalf.eff.pred$depends
#> [1] "mrt.pred"
#> 
#> $thalf.eff.pred$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.iv.obs
#> $thalf.eff.iv.obs$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.iv.obs$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.iv.obs$desc
#> [1] "The effective half-life (as determined from the intravenous MRTobs)"
#> 
#> $thalf.eff.iv.obs$formalsmap
#> $thalf.eff.iv.obs$formalsmap$mrt
#> [1] "mrt.iv.obs"
#> 
#> 
#> $thalf.eff.iv.obs$depends
#> [1] "mrt.iv.obs"
#> 
#> $thalf.eff.iv.obs$datatype
#> [1] "interval"
#> 
#> 
#> $thalf.eff.iv.pred
#> $thalf.eff.iv.pred$FUN
#> [1] "pk.calc.thalf.eff"
#> 
#> $thalf.eff.iv.pred$values
#> [1] FALSE  TRUE
#> 
#> $thalf.eff.iv.pred$desc
#> [1] "The effective half-life (as determined from the intravenous MRTpred)"
#> 
#> $thalf.eff.iv.pred$formalsmap
#> $thalf.eff.iv.pred$formalsmap$mrt
#> [1] "mrt.iv.pred"
#> 
#> 
#> $thalf.eff.iv.pred$depends
#> [1] "mrt.iv.pred"
#> 
#> $thalf.eff.iv.pred$datatype
#> [1] "interval"
#> 
#> 
#> $kel.obs
#> $kel.obs$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.obs$values
#> [1] FALSE  TRUE
#> 
#> $kel.obs$desc
#> [1] "Elimination rate (as calculated from the MRT with observed Clast)"
#> 
#> $kel.obs$formalsmap
#> $kel.obs$formalsmap$mrt
#> [1] "mrt.obs"
#> 
#> 
#> $kel.obs$depends
#> [1] "mrt.obs"
#> 
#> $kel.obs$datatype
#> [1] "interval"
#> 
#> 
#> $kel.pred
#> $kel.pred$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.pred$values
#> [1] FALSE  TRUE
#> 
#> $kel.pred$desc
#> [1] "Elimination rate (as calculated from the MRT with predicted Clast)"
#> 
#> $kel.pred$formalsmap
#> $kel.pred$formalsmap$mrt
#> [1] "mrt.pred"
#> 
#> 
#> $kel.pred$depends
#> [1] "mrt.pred"
#> 
#> $kel.pred$datatype
#> [1] "interval"
#> 
#> 
#> $kel.iv.obs
#> $kel.iv.obs$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.iv.obs$values
#> [1] FALSE  TRUE
#> 
#> $kel.iv.obs$desc
#> [1] "Elimination rate (as calculated from the intravenous MRTobs)"
#> 
#> $kel.iv.obs$formalsmap
#> $kel.iv.obs$formalsmap$mrt
#> [1] "mrt.iv.obs"
#> 
#> 
#> $kel.iv.obs$depends
#> [1] "mrt.iv.obs"
#> 
#> $kel.iv.obs$datatype
#> [1] "interval"
#> 
#> 
#> $kel.iv.pred
#> $kel.iv.pred$FUN
#> [1] "pk.calc.kel"
#> 
#> $kel.iv.pred$values
#> [1] FALSE  TRUE
#> 
#> $kel.iv.pred$desc
#> [1] "Elimination rate (as calculated from the intravenous MRTpred)"
#> 
#> $kel.iv.pred$formalsmap
#> $kel.iv.pred$formalsmap$mrt
#> [1] "mrt.iv.pred"
#> 
#> 
#> $kel.iv.pred$depends
#> [1] "mrt.iv.pred"
#> 
#> $kel.iv.pred$datatype
#> [1] "interval"
#> 
#>