Get the name of the element containing the data for the current object.

Get the name of the element containing the data for the current object.

# S3 method for PKNCAconc
getDataName(object)

# S3 method for PKNCAdata
getDataName(object)

# S3 method for PKNCAdose
getDataName(object)

# S3 method for PKNCAresults
getDataName(object)

getDataName(object)

# S3 method for default
getDataName(object)

Arguments

object
The object to get the data name from.

Value

A character scalar with the name of the data object (or NULL if the method does not apply).

Methods (by class)

default: If no data name exists, returns NULL.

See also