Determine time of maximum observed PK concentration

Input restrictions are:

the conc and time must be the same length,

the time may have no NAs,

NA will be returned if:

the length of conc and time is 0

all conc is 0 or NA

pk.calc.tmax(conc, time, options = list(), first.tmax = NULL, check = TRUE)

Arguments

conc

Concentration measured

time

Time of concentration measurement

options

List of changes to the default PKNCA.options for calculations.

first.tmax

If there is more than time that matches the maximum concentration, should the first be considered as Tmax? If not, then the last is considered Tmax.

check

Run check.conc.time?

Value

the time of the maximum concentration