Mark a formalsmap entry as coming from the reference interval

Used in the formalsmap argument of add.interval.col() to declare that an argument takes the value of param calculated in the reference interval (the interval named by the <parameter>_ref column of the interval specification) rather than in the current interval.

Usage

pknca_ref(param)

is_pknca_ref(x)

Arguments

param

The name of the NCA parameter to take from the reference interval (a single non-missing character string). It does not need to be registered yet when pknca_ref() is called; it is validated when the parameter is calculated.

x

An object to test.

Value

An object of class pknca_ref.

See also

Examples

pknca_ref("aucinf.obs")
#> $param
#> [1] "aucinf.obs"
#> 
#> attr(,"class")
#> [1] "pknca_ref"