Set default options for PKNCA functions

This function will set the default PKNCA options. If given no inputs, it will provide the current option set. If given name/value pairs, it will set the option (as in the options function). If given a name, it will return the value for the parameter. If given the default option as true, it will provide the default options.

PKNCA.options(..., default = FALSE, check = FALSE, name, value)

Arguments

...

options to set or get the value for

default

(re)sets all default options

check

check a single option given, but do not set it (for validation of the values when used in another function)

name

An option name to use with the value.

value

An option value (paired with the name) to set or check (if NULL, ).

Value

If...

no arguments are given

returns the current options.

a value is set (including the defaults)

returns NULL

a single value is requested

the current value of that option is returned as a scalar

multiple values are requested

the current values of those options are returned as a list

Details

Options are either for calculation or summary functions. Calculation options are required for a calculation function to report a result (otherwise the reported value will be NA). Summary options are used during summarization and are used for assessing what values are included in the summary.

See the vignette 'Options for Controlling PKNCA' for a current list of options (vignette("Options-for-Controlling-PKNCA", package="PKNCA")).

See also

Examples

PKNCA.options()
#> $adj.r.squared.factor
#> [1] 1e-04
#> 
#> $max.missing
#> [1] 0.5
#> 
#> $auc.method
#> [1] "lin up/log down"
#> 
#> $conc_above
#> [1] NA
#> 
#> $conc.na
#> [1] "drop"
#> 
#> $conc.blq
#> $conc.blq$first
#> [1] "keep"
#> 
#> $conc.blq$middle
#> [1] "drop"
#> 
#> $conc.blq$last
#> [1] "keep"
#> 
#> 
#> $first.tmax
#> [1] TRUE
#> 
#> $allow.tmax.in.half.life
#> [1] FALSE
#> 
#> $min.hl.points
#> [1] 3
#> 
#> $min.span.ratio
#> [1] 2
#> 
#> $max.aucinf.pext
#> [1] 20
#> 
#> $min.hl.r.squared
#> [1] 0.9
#> 
#> $tau.choices
#> [1] NA
#> 
#> $single.dose.aucs
#>   start end auclast aucall aumclast aumcall aucint.last aucint.last.dose
#> 1     0  24    TRUE  FALSE    FALSE   FALSE       FALSE            FALSE
#> 2     0 Inf   FALSE  FALSE    FALSE   FALSE       FALSE            FALSE
#>   aucint.all aucint.all.dose auclast.dn aucall.dn aumclast.dn aumcall.dn  cmax
#> 1      FALSE           FALSE      FALSE     FALSE       FALSE      FALSE FALSE
#> 2      FALSE           FALSE      FALSE     FALSE       FALSE      FALSE  TRUE
#>    cmin  tmax tlast tfirst clast.obs cl.last cl.all     f mrt.last mrt.iv.last
#> 1 FALSE FALSE FALSE  FALSE     FALSE   FALSE  FALSE FALSE    FALSE       FALSE
#> 2 FALSE  TRUE FALSE  FALSE     FALSE   FALSE  FALSE FALSE    FALSE       FALSE
#>   vss.last vss.iv.last   cav ctrough   ptr  tlag deg.fluc swing  ceoi    ae
#> 1    FALSE       FALSE FALSE   FALSE FALSE FALSE    FALSE FALSE FALSE FALSE
#> 2    FALSE       FALSE FALSE   FALSE FALSE FALSE    FALSE FALSE FALSE FALSE
#>   clr.last clr.obs clr.pred    fe time_above half.life r.squared adj.r.squared
#> 1    FALSE   FALSE    FALSE FALSE      FALSE     FALSE     FALSE         FALSE
#> 2    FALSE   FALSE    FALSE FALSE      FALSE      TRUE     FALSE         FALSE
#>   lambda.z lambda.z.time.first lambda.z.n.points clast.pred span.ratio cmax.dn
#> 1    FALSE               FALSE             FALSE      FALSE      FALSE   FALSE
#> 2    FALSE               FALSE             FALSE      FALSE      FALSE   FALSE
#>   cmin.dn clast.obs.dn clast.pred.dn cav.dn ctrough.dn thalf.eff.last
#> 1   FALSE        FALSE         FALSE  FALSE      FALSE          FALSE
#> 2   FALSE        FALSE         FALSE  FALSE      FALSE          FALSE
#>   thalf.eff.iv.last kel.last kel.iv.last aucinf.obs aucinf.pred aumcinf.obs
#> 1             FALSE    FALSE       FALSE      FALSE       FALSE       FALSE
#> 2             FALSE    FALSE       FALSE       TRUE       FALSE       FALSE
#>   aumcinf.pred aucint.inf.obs aucint.inf.obs.dose aucint.inf.pred
#> 1        FALSE          FALSE               FALSE           FALSE
#> 2        FALSE          FALSE               FALSE           FALSE
#>   aucint.inf.pred.dose aucinf.obs.dn aucinf.pred.dn aumcinf.obs.dn
#> 1                FALSE         FALSE          FALSE          FALSE
#> 2                FALSE         FALSE          FALSE          FALSE
#>   aumcinf.pred.dn aucpext.obs aucpext.pred cl.obs cl.pred mrt.obs mrt.pred
#> 1           FALSE       FALSE        FALSE  FALSE   FALSE   FALSE    FALSE
#> 2           FALSE       FALSE        FALSE  FALSE   FALSE   FALSE    FALSE
#>   mrt.iv.obs mrt.iv.pred mrt.md.obs mrt.md.pred vz.obs vz.pred vss.obs vss.pred
#> 1      FALSE       FALSE      FALSE       FALSE  FALSE   FALSE   FALSE    FALSE
#> 2      FALSE       FALSE      FALSE       FALSE  FALSE   FALSE   FALSE    FALSE
#>   vss.iv.obs vss.iv.pred vss.md.obs vss.md.pred vd.obs vd.pred thalf.eff.obs
#> 1      FALSE       FALSE      FALSE       FALSE  FALSE   FALSE         FALSE
#> 2      FALSE       FALSE      FALSE       FALSE  FALSE   FALSE         FALSE
#>   thalf.eff.pred thalf.eff.iv.obs thalf.eff.iv.pred kel.obs kel.pred kel.iv.obs
#> 1          FALSE            FALSE             FALSE   FALSE    FALSE      FALSE
#> 2          FALSE            FALSE             FALSE   FALSE    FALSE      FALSE
#>   kel.iv.pred
#> 1       FALSE
#> 2       FALSE
#> 
PKNCA.options(default=TRUE)
PKNCA.options("auc.method")
#> [1] "lin up/log down"
PKNCA.options(name="auc.method")
#> [1] "lin up/log down"
PKNCA.options(auc.method="lin up/log down", min.hl.points=3)