Methods for imputation of data with PKNCA

Methods for imputation of data with PKNCA

PKNCA_impute_method_start_conc0(conc, time, start = 0, ..., options = list())

PKNCA_impute_method_start_cmin(conc, time, start, end, ..., options = list())

PKNCA_impute_method_start_predose(
  conc,
  time,
  start,
  end,
  ...,
  max_shift = NA_real_,
  options = list()
)

Arguments

conc

Concentration measured

time

Time of concentration measurement (must be monotonically increasing and the same length as the concentration data)

start, end

The start and end of the interval

...

ignored

options

List of changes to the default PKNCA.options for calculations.

max_shift

The maximum amount of time to shift a concentration forward (defaults to 5% of the interval duration, i.e. 0.05*(end - start))

Value

A data.frame with one column named conc with imputed concentrations and one column named time with the times.

Functions

PKNCA_impute_method_start_conc0(): Add a new concentration of 0 at the start time, even if a nonzero concentration exists at that time (usually used with single-dose data)

PKNCA_impute_method_start_cmin(): Add a new concentration of the minimum during the interval at the start time (usually used with multiple-dose data)

PKNCA_impute_method_start_predose(): Shift a predose concentration to become the time zero concentration (only if a time zero concentration does not exist)