Create a PKNCAdata object.

PKNCAdata combines PKNCAconc and PKNCAdose and adds in the intervals for PK calculations.

PKNCAdata(data.conc, data.dose, ...)

# S3 method for PKNCAconc
PKNCAdata(data.conc, data.dose, ...)

# S3 method for PKNCAdose
PKNCAdata(data.conc, data.dose, ...)

# S3 method for default
PKNCAdata(
  data.conc,
  data.dose,
  ...,
  formula.conc,
  formula.dose,
  intervals,
  options = list()
)

Arguments

data.conc
Concentration data as a PKNCAconc object or a data frame

data.dose
Dosing data as a PKNCAdose object (see details)

...
arguments passed to PKNCAdata.default

formula.conc
Formula for making a PKNCAconc object with data.conc. This must be given if data.conc is a data.frame, and it must not be given if data.conc is a PKNCAconc object.

formula.dose
Formula for making a PKNCAdose object with data.dose. This must be given if data.dose is a data.frame, and it must not be given if data.dose is a PKNCAdose object.

intervals
A data frame with the AUC interval specifications as defined in check.interval.specification. If missing, this will be automatically chosen by choose.auc.intervals. (see details)

options
List of changes to the default PKNCA.options for calculations.

Value

A PKNCAdata object with concentration, dose, interval, and calculation options stored (note that PKNCAdata objects can also have results after a NCA calculations are done to the data).

Details

If data.dose is not given or is NA, then the intervals must be given. At least one of data.dose and intervals must be given.

See also