Choose intervals to compute AUCs from time and dosing information

Intervals for AUC are selected by the following metrics:

If only one dose is administered, use the PKNCA.options("single.dose.aucs")

If more than one dose is administered, estimate the AUC between any two doses that have PK taken at both of the dosing times and at least one time between the doses.

For the final dose of multiple doses, try to determine the dosing interval (\(\tau\)) and estimate the AUC in that interval if multiple samples are taken in the interval.

If there are samples \(> \tau\) after the last dose, calculate the half life after the last dose.

choose.auc.intervals(
  time.conc,
  time.dosing,
  options = list(),
  single.dose.aucs = NULL
)

Arguments

time.conc

Time of concentration measurement

time.dosing

Time of dosing

options

List of changes to the default PKNCA.options for calculations.

single.dose.aucs

The AUC specification for single dosing.

Value

A data frame with columns for start, end, auc.type, and half.life. See check.interval.specification for column definitions. The data frame may have zero rows if no intervals could be found.

See also