Get the impute function from either the intervals column or from the method

Get the impute function from either the intervals column or from the method

Usage

get_impute_method(intervals, impute)

Arguments

intervals

the data.frame of intervals

impute

the imputation definition

Value

The imputation function vector