Compute NCA for multiple intervals

Compute NCA for multiple intervals

pk.nca.intervals(
  data_conc,
  data_dose,
  data_intervals,
  options,
  verbose = FALSE
)

Arguments

data_conc
A data.frame or tibble with standardized column names as output from prepare_PKNCAconc()

data_dose
A data.frame or tibble with standardized column names as output from prepare_PKNCAdose()

data_intervals
A data.frame or tibble with standardized column names as output from prepare_PKNCAintervals()

options
List of changes to the default PKNCA.options for calculations.

verbose
Indicate, by message(), the current state of calculation.

Value

A data.frame with all NCA results