Create a unit assignment and conversion table

This data.frame is typically used for the units argument for PKNCAdata().

pknca_units_table(concu, doseu, amountu, timeu, conversions = data.frame())

Arguments

concu, doseu, amountu, timeu

Units for concentration, dose, amount, and time

conversions

An optional data.frame with columns of c("PPORRESU", "PPSTRESU", "conversion_factor") for the original calculation units, the standardized units, and a conversion factor to multiply the initial value by to get a standardized value.

Value

A unit conversion table with columns for "PPTESTCD" and "PPORRESU" if

conversions is not given, and adding "PPSTRESU" and "conversion_factor" if conversions is given.

See also

Examples

pknca_units_table() # only parameters that are unitless
#>    PPORRESU          PPTESTCD
#> 1  unitless         r.squared
#> 2  unitless     adj.r.squared
#> 3  fraction                 f
#> 4  fraction               ptr
#> 5  fraction                fe
#> 6  fraction        span.ratio
#> 7         %          deg.fluc
#> 8         %             swing
#> 9         %       aucpext.obs
#> 10        %      aucpext.pred
#> 11    count lambda.z.n.points
pknca_units_table(
  concu="ng/mL", doseu="mg/kg", amountu="mg", timeu="hr"
)
#>                PPORRESU             PPTESTCD
#> 1              unitless            r.squared
#> 2              unitless        adj.r.squared
#> 3              fraction                    f
#> 4              fraction                  ptr
#> 5              fraction                   fe
#> 6              fraction           span.ratio
#> 7                     %             deg.fluc
#> 8                     %                swing
#> 9                     %          aucpext.obs
#> 10                    %         aucpext.pred
#> 11                count    lambda.z.n.points
#> 12                   hr                start
#> 13                   hr                  end
#> 14                   hr                 tmax
#> 15                   hr                tlast
#> 16                   hr               tfirst
#> 17                   hr             mrt.last
#> 18                   hr          mrt.iv.last
#> 19                   hr                 tlag
#> 20                   hr           time_above
#> 21                   hr            half.life
#> 22                   hr  lambda.z.time.first
#> 23                   hr       thalf.eff.last
#> 24                   hr    thalf.eff.iv.last
#> 25                   hr              mrt.obs
#> 26                   hr             mrt.pred
#> 27                   hr           mrt.iv.obs
#> 28                   hr          mrt.iv.pred
#> 29                   hr           mrt.md.obs
#> 30                   hr          mrt.md.pred
#> 31                   hr        thalf.eff.obs
#> 32                   hr       thalf.eff.pred
#> 33                   hr     thalf.eff.iv.obs
#> 34                   hr    thalf.eff.iv.pred
#> 35                 1/hr             lambda.z
#> 36                 1/hr             kel.last
#> 37                 1/hr          kel.iv.last
#> 38                 1/hr              kel.obs
#> 39                 1/hr             kel.pred
#> 40                 1/hr           kel.iv.obs
#> 41                 1/hr          kel.iv.pred
#> 42                ng/mL                 cmax
#> 43                ng/mL                 cmin
#> 44                ng/mL            clast.obs
#> 45                ng/mL                  cav
#> 46                ng/mL              ctrough
#> 47                ng/mL                 ceoi
#> 48                ng/mL           clast.pred
#> 49                   mg                   ae
#> 50      (ng/mL)/(mg/kg)              cmax.dn
#> 51      (ng/mL)/(mg/kg)              cmin.dn
#> 52      (ng/mL)/(mg/kg)         clast.obs.dn
#> 53      (ng/mL)/(mg/kg)        clast.pred.dn
#> 54      (ng/mL)/(mg/kg)               cav.dn
#> 55      (ng/mL)/(mg/kg)           ctrough.dn
#> 56      (mg/kg)/(ng/mL)             vss.last
#> 57      (mg/kg)/(ng/mL)          vss.iv.last
#> 58      (mg/kg)/(ng/mL)               vz.obs
#> 59      (mg/kg)/(ng/mL)              vz.pred
#> 60      (mg/kg)/(ng/mL)              vss.obs
#> 61      (mg/kg)/(ng/mL)             vss.pred
#> 62      (mg/kg)/(ng/mL)           vss.iv.obs
#> 63      (mg/kg)/(ng/mL)          vss.iv.pred
#> 64      (mg/kg)/(ng/mL)           vss.md.obs
#> 65      (mg/kg)/(ng/mL)          vss.md.pred
#> 66      (mg/kg)/(ng/mL)               vd.obs
#> 67      (mg/kg)/(ng/mL)              vd.pred
#> 68             hr*ng/mL              auclast
#> 69             hr*ng/mL               aucall
#> 70             hr*ng/mL          aucint.last
#> 71             hr*ng/mL     aucint.last.dose
#> 72             hr*ng/mL           aucint.all
#> 73             hr*ng/mL      aucint.all.dose
#> 74             hr*ng/mL       sparse_auclast
#> 75             hr*ng/mL           aucinf.obs
#> 76             hr*ng/mL          aucinf.pred
#> 77             hr*ng/mL       aucint.inf.obs
#> 78             hr*ng/mL  aucint.inf.obs.dose
#> 79             hr*ng/mL      aucint.inf.pred
#> 80             hr*ng/mL aucint.inf.pred.dose
#> 81           hr^2*ng/mL             aumclast
#> 82           hr^2*ng/mL              aumcall
#> 83           hr^2*ng/mL          aumcinf.obs
#> 84           hr^2*ng/mL         aumcinf.pred
#> 85   (hr*ng/mL)/(mg/kg)           auclast.dn
#> 86   (hr*ng/mL)/(mg/kg)            aucall.dn
#> 87   (hr*ng/mL)/(mg/kg)        aucinf.obs.dn
#> 88   (hr*ng/mL)/(mg/kg)       aucinf.pred.dn
#> 89 (hr^2*ng/mL)/(mg/kg)          aumclast.dn
#> 90 (hr^2*ng/mL)/(mg/kg)           aumcall.dn
#> 91 (hr^2*ng/mL)/(mg/kg)       aumcinf.obs.dn
#> 92 (hr^2*ng/mL)/(mg/kg)      aumcinf.pred.dn
#> 93   (mg/kg)/(hr*ng/mL)              cl.last
#> 94   (mg/kg)/(hr*ng/mL)               cl.all
#> 95   (mg/kg)/(hr*ng/mL)               cl.obs
#> 96   (mg/kg)/(hr*ng/mL)              cl.pred
#> 97        mg/(hr*ng/mL)             clr.last
#> 98        mg/(hr*ng/mL)              clr.obs
#> 99        mg/(hr*ng/mL)             clr.pred
pknca_units_table(
  concu="ng/mL", doseu="mg/kg", amountu="mg", timeu="hr",
  # Convert clearance and volume units to more understandable units with
  # automatic unit conversion
  conversions=data.frame(
    PPORRESU=c("(mg/kg)/(hr*ng/mL)", "(mg/kg)/(ng/mL)"),
    PPSTRESU=c("mL/hr/kg", "mL/kg")
  )
)
#>                PPORRESU             PPTESTCD             PPSTRESU
#> 1              unitless            r.squared             unitless
#> 2              unitless        adj.r.squared             unitless
#> 3              fraction                    f             fraction
#> 4              fraction                  ptr             fraction
#> 5              fraction                   fe             fraction
#> 6              fraction           span.ratio             fraction
#> 7                     %             deg.fluc                    %
#> 8                     %                swing                    %
#> 9                     %          aucpext.obs                    %
#> 10                    %         aucpext.pred                    %
#> 11                count    lambda.z.n.points                count
#> 12                   hr                start                   hr
#> 13                   hr                  end                   hr
#> 14                   hr                 tmax                   hr
#> 15                   hr                tlast                   hr
#> 16                   hr               tfirst                   hr
#> 17                   hr             mrt.last                   hr
#> 18                   hr          mrt.iv.last                   hr
#> 19                   hr                 tlag                   hr
#> 20                   hr           time_above                   hr
#> 21                   hr            half.life                   hr
#> 22                   hr  lambda.z.time.first                   hr
#> 23                   hr       thalf.eff.last                   hr
#> 24                   hr    thalf.eff.iv.last                   hr
#> 25                   hr              mrt.obs                   hr
#> 26                   hr             mrt.pred                   hr
#> 27                   hr           mrt.iv.obs                   hr
#> 28                   hr          mrt.iv.pred                   hr
#> 29                   hr           mrt.md.obs                   hr
#> 30                   hr          mrt.md.pred                   hr
#> 31                   hr        thalf.eff.obs                   hr
#> 32                   hr       thalf.eff.pred                   hr
#> 33                   hr     thalf.eff.iv.obs                   hr
#> 34                   hr    thalf.eff.iv.pred                   hr
#> 35                 1/hr             lambda.z                 1/hr
#> 36                 1/hr             kel.last                 1/hr
#> 37                 1/hr          kel.iv.last                 1/hr
#> 38                 1/hr              kel.obs                 1/hr
#> 39                 1/hr             kel.pred                 1/hr
#> 40                 1/hr           kel.iv.obs                 1/hr
#> 41                 1/hr          kel.iv.pred                 1/hr
#> 42                ng/mL                 cmax                ng/mL
#> 43                ng/mL                 cmin                ng/mL
#> 44                ng/mL            clast.obs                ng/mL
#> 45                ng/mL                  cav                ng/mL
#> 46                ng/mL              ctrough                ng/mL
#> 47                ng/mL                 ceoi                ng/mL
#> 48                ng/mL           clast.pred                ng/mL
#> 49                   mg                   ae                   mg
#> 50      (ng/mL)/(mg/kg)              cmax.dn      (ng/mL)/(mg/kg)
#> 51      (ng/mL)/(mg/kg)              cmin.dn      (ng/mL)/(mg/kg)
#> 52      (ng/mL)/(mg/kg)         clast.obs.dn      (ng/mL)/(mg/kg)
#> 53      (ng/mL)/(mg/kg)        clast.pred.dn      (ng/mL)/(mg/kg)
#> 54      (ng/mL)/(mg/kg)               cav.dn      (ng/mL)/(mg/kg)
#> 55      (ng/mL)/(mg/kg)           ctrough.dn      (ng/mL)/(mg/kg)
#> 56      (mg/kg)/(ng/mL)             vss.last                mL/kg
#> 57      (mg/kg)/(ng/mL)          vss.iv.last                mL/kg
#> 58      (mg/kg)/(ng/mL)               vz.obs                mL/kg
#> 59      (mg/kg)/(ng/mL)              vz.pred                mL/kg
#> 60      (mg/kg)/(ng/mL)              vss.obs                mL/kg
#> 61      (mg/kg)/(ng/mL)             vss.pred                mL/kg
#> 62      (mg/kg)/(ng/mL)           vss.iv.obs                mL/kg
#> 63      (mg/kg)/(ng/mL)          vss.iv.pred                mL/kg
#> 64      (mg/kg)/(ng/mL)           vss.md.obs                mL/kg
#> 65      (mg/kg)/(ng/mL)          vss.md.pred                mL/kg
#> 66      (mg/kg)/(ng/mL)               vd.obs                mL/kg
#> 67      (mg/kg)/(ng/mL)              vd.pred                mL/kg
#> 68             hr*ng/mL              auclast             hr*ng/mL
#> 69             hr*ng/mL               aucall             hr*ng/mL
#> 70             hr*ng/mL          aucint.last             hr*ng/mL
#> 71             hr*ng/mL     aucint.last.dose             hr*ng/mL
#> 72             hr*ng/mL           aucint.all             hr*ng/mL
#> 73             hr*ng/mL      aucint.all.dose             hr*ng/mL
#> 74             hr*ng/mL       sparse_auclast             hr*ng/mL
#> 75             hr*ng/mL           aucinf.obs             hr*ng/mL
#> 76             hr*ng/mL          aucinf.pred             hr*ng/mL
#> 77             hr*ng/mL       aucint.inf.obs             hr*ng/mL
#> 78             hr*ng/mL  aucint.inf.obs.dose             hr*ng/mL
#> 79             hr*ng/mL      aucint.inf.pred             hr*ng/mL
#> 80             hr*ng/mL aucint.inf.pred.dose             hr*ng/mL
#> 81           hr^2*ng/mL             aumclast           hr^2*ng/mL
#> 82           hr^2*ng/mL              aumcall           hr^2*ng/mL
#> 83           hr^2*ng/mL          aumcinf.obs           hr^2*ng/mL
#> 84           hr^2*ng/mL         aumcinf.pred           hr^2*ng/mL
#> 85   (hr*ng/mL)/(mg/kg)           auclast.dn   (hr*ng/mL)/(mg/kg)
#> 86   (hr*ng/mL)/(mg/kg)            aucall.dn   (hr*ng/mL)/(mg/kg)
#> 87   (hr*ng/mL)/(mg/kg)        aucinf.obs.dn   (hr*ng/mL)/(mg/kg)
#> 88   (hr*ng/mL)/(mg/kg)       aucinf.pred.dn   (hr*ng/mL)/(mg/kg)
#> 89 (hr^2*ng/mL)/(mg/kg)          aumclast.dn (hr^2*ng/mL)/(mg/kg)
#> 90 (hr^2*ng/mL)/(mg/kg)           aumcall.dn (hr^2*ng/mL)/(mg/kg)
#> 91 (hr^2*ng/mL)/(mg/kg)       aumcinf.obs.dn (hr^2*ng/mL)/(mg/kg)
#> 92 (hr^2*ng/mL)/(mg/kg)      aumcinf.pred.dn (hr^2*ng/mL)/(mg/kg)
#> 93   (mg/kg)/(hr*ng/mL)              cl.last             mL/hr/kg
#> 94   (mg/kg)/(hr*ng/mL)               cl.all             mL/hr/kg
#> 95   (mg/kg)/(hr*ng/mL)               cl.obs             mL/hr/kg
#> 96   (mg/kg)/(hr*ng/mL)              cl.pred             mL/hr/kg
#> 97        mg/(hr*ng/mL)             clr.last        mg/(hr*ng/mL)
#> 98        mg/(hr*ng/mL)              clr.obs        mg/(hr*ng/mL)
#> 99        mg/(hr*ng/mL)             clr.pred        mg/(hr*ng/mL)
#>    conversion_factor
#> 1              1e+00
#> 2              1e+00
#> 3              1e+00
#> 4              1e+00
#> 5              1e+00
#> 6              1e+00
#> 7              1e+00
#> 8              1e+00
#> 9              1e+00
#> 10             1e+00
#> 11             1e+00
#> 12             1e+00
#> 13             1e+00
#> 14             1e+00
#> 15             1e+00
#> 16             1e+00
#> 17             1e+00
#> 18             1e+00
#> 19             1e+00
#> 20             1e+00
#> 21             1e+00
#> 22             1e+00
#> 23             1e+00
#> 24             1e+00
#> 25             1e+00
#> 26             1e+00
#> 27             1e+00
#> 28             1e+00
#> 29             1e+00
#> 30             1e+00
#> 31             1e+00
#> 32             1e+00
#> 33             1e+00
#> 34             1e+00
#> 35             1e+00
#> 36             1e+00
#> 37             1e+00
#> 38             1e+00
#> 39             1e+00
#> 40             1e+00
#> 41             1e+00
#> 42             1e+00
#> 43             1e+00
#> 44             1e+00
#> 45             1e+00
#> 46             1e+00
#> 47             1e+00
#> 48             1e+00
#> 49             1e+00
#> 50             1e+00
#> 51             1e+00
#> 52             1e+00
#> 53             1e+00
#> 54             1e+00
#> 55             1e+00
#> 56             1e+06
#> 57             1e+06
#> 58             1e+06
#> 59             1e+06
#> 60             1e+06
#> 61             1e+06
#> 62             1e+06
#> 63             1e+06
#> 64             1e+06
#> 65             1e+06
#> 66             1e+06
#> 67             1e+06
#> 68             1e+00
#> 69             1e+00
#> 70             1e+00
#> 71             1e+00
#> 72             1e+00
#> 73             1e+00
#> 74             1e+00
#> 75             1e+00
#> 76             1e+00
#> 77             1e+00
#> 78             1e+00
#> 79             1e+00
#> 80             1e+00
#> 81             1e+00
#> 82             1e+00
#> 83             1e+00
#> 84             1e+00
#> 85             1e+00
#> 86             1e+00
#> 87             1e+00
#> 88             1e+00
#> 89             1e+00
#> 90             1e+00
#> 91             1e+00
#> 92             1e+00
#> 93             1e+06
#> 94             1e+06
#> 95             1e+06
#> 96             1e+06
#> 97             1e+00
#> 98             1e+00
#> 99             1e+00
pknca_units_table(
  concu="mg/L", doseu="mg/kg", amountu="mg", timeu="hr",
  # Convert clearance and volume units to molar units (assuming 
  conversions=data.frame(
    PPORRESU=c("mg/L", "(mg/kg)/(hr*ng/mL)", "(mg/kg)/(ng/mL)"),
    PPSTRESU=c("mmol/L", "mL/hr/kg", "mL/kg"),
    # Manual conversion of concentration units from ng/mL to mmol/L (assuming
    # a molecular weight of 138.121 g/mol)
    conversion_factor=c(1/138.121, NA, NA)
  )
)
#> Warning: The following unit conversions were supplied but do not match any units to convert: '(mg/kg)/(hr*ng/mL)', '(mg/kg)/(ng/mL)'
#>               PPORRESU             PPTESTCD            PPSTRESU
#> 1             unitless            r.squared            unitless
#> 2             unitless        adj.r.squared            unitless
#> 3             fraction                    f            fraction
#> 4             fraction                  ptr            fraction
#> 5             fraction                   fe            fraction
#> 6             fraction           span.ratio            fraction
#> 7                    %             deg.fluc                   %
#> 8                    %                swing                   %
#> 9                    %          aucpext.obs                   %
#> 10                   %         aucpext.pred                   %
#> 11               count    lambda.z.n.points               count
#> 12                  hr                start                  hr
#> 13                  hr                  end                  hr
#> 14                  hr                 tmax                  hr
#> 15                  hr                tlast                  hr
#> 16                  hr               tfirst                  hr
#> 17                  hr             mrt.last                  hr
#> 18                  hr          mrt.iv.last                  hr
#> 19                  hr                 tlag                  hr
#> 20                  hr           time_above                  hr
#> 21                  hr            half.life                  hr
#> 22                  hr  lambda.z.time.first                  hr
#> 23                  hr       thalf.eff.last                  hr
#> 24                  hr    thalf.eff.iv.last                  hr
#> 25                  hr              mrt.obs                  hr
#> 26                  hr             mrt.pred                  hr
#> 27                  hr           mrt.iv.obs                  hr
#> 28                  hr          mrt.iv.pred                  hr
#> 29                  hr           mrt.md.obs                  hr
#> 30                  hr          mrt.md.pred                  hr
#> 31                  hr        thalf.eff.obs                  hr
#> 32                  hr       thalf.eff.pred                  hr
#> 33                  hr     thalf.eff.iv.obs                  hr
#> 34                  hr    thalf.eff.iv.pred                  hr
#> 35                1/hr             lambda.z                1/hr
#> 36                1/hr             kel.last                1/hr
#> 37                1/hr          kel.iv.last                1/hr
#> 38                1/hr              kel.obs                1/hr
#> 39                1/hr             kel.pred                1/hr
#> 40                1/hr           kel.iv.obs                1/hr
#> 41                1/hr          kel.iv.pred                1/hr
#> 42                mg/L                 cmax              mmol/L
#> 43                mg/L                 cmin              mmol/L
#> 44                mg/L            clast.obs              mmol/L
#> 45                mg/L                  cav              mmol/L
#> 46                mg/L              ctrough              mmol/L
#> 47                mg/L                 ceoi              mmol/L
#> 48                mg/L           clast.pred              mmol/L
#> 49                  mg                   ae                  mg
#> 50      (mg/L)/(mg/kg)              cmax.dn      (mg/L)/(mg/kg)
#> 51      (mg/L)/(mg/kg)              cmin.dn      (mg/L)/(mg/kg)
#> 52      (mg/L)/(mg/kg)         clast.obs.dn      (mg/L)/(mg/kg)
#> 53      (mg/L)/(mg/kg)        clast.pred.dn      (mg/L)/(mg/kg)
#> 54      (mg/L)/(mg/kg)               cav.dn      (mg/L)/(mg/kg)
#> 55      (mg/L)/(mg/kg)           ctrough.dn      (mg/L)/(mg/kg)
#> 56      (mg/kg)/(mg/L)             vss.last      (mg/kg)/(mg/L)
#> 57      (mg/kg)/(mg/L)          vss.iv.last      (mg/kg)/(mg/L)
#> 58      (mg/kg)/(mg/L)               vz.obs      (mg/kg)/(mg/L)
#> 59      (mg/kg)/(mg/L)              vz.pred      (mg/kg)/(mg/L)
#> 60      (mg/kg)/(mg/L)              vss.obs      (mg/kg)/(mg/L)
#> 61      (mg/kg)/(mg/L)             vss.pred      (mg/kg)/(mg/L)
#> 62      (mg/kg)/(mg/L)           vss.iv.obs      (mg/kg)/(mg/L)
#> 63      (mg/kg)/(mg/L)          vss.iv.pred      (mg/kg)/(mg/L)
#> 64      (mg/kg)/(mg/L)           vss.md.obs      (mg/kg)/(mg/L)
#> 65      (mg/kg)/(mg/L)          vss.md.pred      (mg/kg)/(mg/L)
#> 66      (mg/kg)/(mg/L)               vd.obs      (mg/kg)/(mg/L)
#> 67      (mg/kg)/(mg/L)              vd.pred      (mg/kg)/(mg/L)
#> 68             hr*mg/L              auclast             hr*mg/L
#> 69             hr*mg/L               aucall             hr*mg/L
#> 70             hr*mg/L          aucint.last             hr*mg/L
#> 71             hr*mg/L     aucint.last.dose             hr*mg/L
#> 72             hr*mg/L           aucint.all             hr*mg/L
#> 73             hr*mg/L      aucint.all.dose             hr*mg/L
#> 74             hr*mg/L       sparse_auclast             hr*mg/L
#> 75             hr*mg/L           aucinf.obs             hr*mg/L
#> 76             hr*mg/L          aucinf.pred             hr*mg/L
#> 77             hr*mg/L       aucint.inf.obs             hr*mg/L
#> 78             hr*mg/L  aucint.inf.obs.dose             hr*mg/L
#> 79             hr*mg/L      aucint.inf.pred             hr*mg/L
#> 80             hr*mg/L aucint.inf.pred.dose             hr*mg/L
#> 81           hr^2*mg/L             aumclast           hr^2*mg/L
#> 82           hr^2*mg/L              aumcall           hr^2*mg/L
#> 83           hr^2*mg/L          aumcinf.obs           hr^2*mg/L
#> 84           hr^2*mg/L         aumcinf.pred           hr^2*mg/L
#> 85   (hr*mg/L)/(mg/kg)           auclast.dn   (hr*mg/L)/(mg/kg)
#> 86   (hr*mg/L)/(mg/kg)            aucall.dn   (hr*mg/L)/(mg/kg)
#> 87   (hr*mg/L)/(mg/kg)        aucinf.obs.dn   (hr*mg/L)/(mg/kg)
#> 88   (hr*mg/L)/(mg/kg)       aucinf.pred.dn   (hr*mg/L)/(mg/kg)
#> 89 (hr^2*mg/L)/(mg/kg)          aumclast.dn (hr^2*mg/L)/(mg/kg)
#> 90 (hr^2*mg/L)/(mg/kg)           aumcall.dn (hr^2*mg/L)/(mg/kg)
#> 91 (hr^2*mg/L)/(mg/kg)       aumcinf.obs.dn (hr^2*mg/L)/(mg/kg)
#> 92 (hr^2*mg/L)/(mg/kg)      aumcinf.pred.dn (hr^2*mg/L)/(mg/kg)
#> 93   (mg/kg)/(hr*mg/L)              cl.last   (mg/kg)/(hr*mg/L)
#> 94   (mg/kg)/(hr*mg/L)               cl.all   (mg/kg)/(hr*mg/L)
#> 95   (mg/kg)/(hr*mg/L)               cl.obs   (mg/kg)/(hr*mg/L)
#> 96   (mg/kg)/(hr*mg/L)              cl.pred   (mg/kg)/(hr*mg/L)
#> 97        mg/(hr*mg/L)             clr.last        mg/(hr*mg/L)
#> 98        mg/(hr*mg/L)              clr.obs        mg/(hr*mg/L)
#> 99        mg/(hr*mg/L)             clr.pred        mg/(hr*mg/L)
#>    conversion_factor
#> 1        1.000000000
#> 2        1.000000000
#> 3        1.000000000
#> 4        1.000000000
#> 5        1.000000000
#> 6        1.000000000
#> 7        1.000000000
#> 8        1.000000000
#> 9        1.000000000
#> 10       1.000000000
#> 11       1.000000000
#> 12       1.000000000
#> 13       1.000000000
#> 14       1.000000000
#> 15       1.000000000
#> 16       1.000000000
#> 17       1.000000000
#> 18       1.000000000
#> 19       1.000000000
#> 20       1.000000000
#> 21       1.000000000
#> 22       1.000000000
#> 23       1.000000000
#> 24       1.000000000
#> 25       1.000000000
#> 26       1.000000000
#> 27       1.000000000
#> 28       1.000000000
#> 29       1.000000000
#> 30       1.000000000
#> 31       1.000000000
#> 32       1.000000000
#> 33       1.000000000
#> 34       1.000000000
#> 35       1.000000000
#> 36       1.000000000
#> 37       1.000000000
#> 38       1.000000000
#> 39       1.000000000
#> 40       1.000000000
#> 41       1.000000000
#> 42       0.007240029
#> 43       0.007240029
#> 44       0.007240029
#> 45       0.007240029
#> 46       0.007240029
#> 47       0.007240029
#> 48       0.007240029
#> 49       1.000000000
#> 50       1.000000000
#> 51       1.000000000
#> 52       1.000000000
#> 53       1.000000000
#> 54       1.000000000
#> 55       1.000000000
#> 56       1.000000000
#> 57       1.000000000
#> 58       1.000000000
#> 59       1.000000000
#> 60       1.000000000
#> 61       1.000000000
#> 62       1.000000000
#> 63       1.000000000
#> 64       1.000000000
#> 65       1.000000000
#> 66       1.000000000
#> 67       1.000000000
#> 68       1.000000000
#> 69       1.000000000
#> 70       1.000000000
#> 71       1.000000000
#> 72       1.000000000
#> 73       1.000000000
#> 74       1.000000000
#> 75       1.000000000
#> 76       1.000000000
#> 77       1.000000000
#> 78       1.000000000
#> 79       1.000000000
#> 80       1.000000000
#> 81       1.000000000
#> 82       1.000000000
#> 83       1.000000000
#> 84       1.000000000
#> 85       1.000000000
#> 86       1.000000000
#> 87       1.000000000
#> 88       1.000000000
#> 89       1.000000000
#> 90       1.000000000
#> 91       1.000000000
#> 92       1.000000000
#> 93       1.000000000
#> 94       1.000000000
#> 95       1.000000000
#> 96       1.000000000
#> 97       1.000000000
#> 98       1.000000000
#> 99       1.000000000