Create a unit assignment and conversion table

This data.frame is typically used for the units argument for PKNCAdata(). If a unit is not given, then all of the units derived from that unit will be NA.

Usage

pknca_units_table(concu, doseu, amountu, timeu, conversions = data.frame())

Arguments

concu, doseu, amountu, timeu

Units for concentration, dose, amount, and time

conversions

An optional data.frame with columns of c("PPORRESU", "PPSTRESU", "conversion_factor") for the original calculation units, the standardized units, and a conversion factor to multiply the initial value by to get a standardized value.

Value

A unit conversion table with columns for "PPTESTCD" and "PPORRESU" if

conversions is not given, and adding "PPSTRESU" and "conversion_factor" if conversions is given.

See also

Examples

pknca_units_table() # only parameters that are unitless
#>     PPORRESU             PPTESTCD
#> 1   unitless            r.squared
#> 2   unitless        adj.r.squared
#> 3   fraction                    f
#> 4   fraction                  ptr
#> 5   fraction                   fe
#> 6   fraction           span.ratio
#> 7          %             deg.fluc
#> 8          %                swing
#> 9          %       aucivpbextlast
#> 10         %        aucivpbextall
#> 11         %   aucivpbextint.last
#> 12         %    aucivpbextint.all
#> 13         %    aucivpbextinf.obs
#> 14         %   aucivpbextinf.pred
#> 15         %          aucpext.obs
#> 16         %         aucpext.pred
#> 17     count           count_conc
#> 18     count    lambda.z.n.points
#> 19      <NA>                start
#> 20      <NA>                  end
#> 21      <NA>                 tmax
#> 22      <NA>                tlast
#> 23      <NA>               tfirst
#> 24      <NA>             mrt.last
#> 25      <NA>          mrt.iv.last
#> 26      <NA>                 tlag
#> 27      <NA>           time_above
#> 28      <NA>            half.life
#> 29      <NA>  lambda.z.time.first
#> 30      <NA>       thalf.eff.last
#> 31      <NA>    thalf.eff.iv.last
#> 32      <NA>              mrt.obs
#> 33      <NA>             mrt.pred
#> 34      <NA>           mrt.iv.obs
#> 35      <NA>          mrt.iv.pred
#> 36      <NA>           mrt.md.obs
#> 37      <NA>          mrt.md.pred
#> 38      <NA>        thalf.eff.obs
#> 39      <NA>       thalf.eff.pred
#> 40      <NA>     thalf.eff.iv.obs
#> 41      <NA>    thalf.eff.iv.pred
#> 42      <NA>             lambda.z
#> 43      <NA>             kel.last
#> 44      <NA>          kel.iv.last
#> 45      <NA>              kel.obs
#> 46      <NA>             kel.pred
#> 47      <NA>           kel.iv.obs
#> 48      <NA>          kel.iv.pred
#> 49      <NA>                   c0
#> 50      <NA>                 cmax
#> 51      <NA>                 cmin
#> 52      <NA>            clast.obs
#> 53      <NA>                  cav
#> 54      <NA>              ctrough
#> 55      <NA>               cstart
#> 56      <NA>                 ceoi
#> 57      <NA>           clast.pred
#> 58      <NA>                   ae
#> 59      <NA>              cmax.dn
#> 60      <NA>              cmin.dn
#> 61      <NA>         clast.obs.dn
#> 62      <NA>        clast.pred.dn
#> 63      <NA>               cav.dn
#> 64      <NA>           ctrough.dn
#> 65      <NA>             vss.last
#> 66      <NA>          vss.iv.last
#> 67      <NA>               vz.obs
#> 68      <NA>              vz.pred
#> 69      <NA>              vss.obs
#> 70      <NA>             vss.pred
#> 71      <NA>           vss.iv.obs
#> 72      <NA>          vss.iv.pred
#> 73      <NA>           vss.md.obs
#> 74      <NA>          vss.md.pred
#> 75      <NA>              auclast
#> 76      <NA>               aucall
#> 77      <NA>          aucint.last
#> 78      <NA>     aucint.last.dose
#> 79      <NA>           aucint.all
#> 80      <NA>      aucint.all.dose
#> 81      <NA> aucabove.predose.all
#> 82      <NA>  aucabove.trough.all
#> 83      <NA>       sparse_auclast
#> 84      <NA>            aucivlast
#> 85      <NA>             aucivall
#> 86      <NA>        aucivint.last
#> 87      <NA>         aucivint.all
#> 88      <NA>           aucinf.obs
#> 89      <NA>          aucinf.pred
#> 90      <NA>       aucint.inf.obs
#> 91      <NA>  aucint.inf.obs.dose
#> 92      <NA>      aucint.inf.pred
#> 93      <NA> aucint.inf.pred.dose
#> 94      <NA>         aucivinf.obs
#> 95      <NA>        aucivinf.pred
#> 96      <NA>             aumclast
#> 97      <NA>              aumcall
#> 98      <NA>          aumcinf.obs
#> 99      <NA>         aumcinf.pred
#> 100     <NA>           auclast.dn
#> 101     <NA>            aucall.dn
#> 102     <NA>        aucinf.obs.dn
#> 103     <NA>       aucinf.pred.dn
#> 104     <NA>          aumclast.dn
#> 105     <NA>           aumcall.dn
#> 106     <NA>       aumcinf.obs.dn
#> 107     <NA>      aumcinf.pred.dn
#> 108     <NA>              cl.last
#> 109     <NA>               cl.all
#> 110     <NA>               cl.obs
#> 111     <NA>              cl.pred
#> 112     <NA>             clr.last
#> 113     <NA>              clr.obs
#> 114     <NA>             clr.pred
pknca_units_table(
  concu="ng/mL", doseu="mg/kg", amountu="mg", timeu="hr"
)
#>                 PPORRESU             PPTESTCD
#> 1               unitless            r.squared
#> 2               unitless        adj.r.squared
#> 3               fraction                    f
#> 4               fraction                  ptr
#> 5               fraction                   fe
#> 6               fraction           span.ratio
#> 7                      %             deg.fluc
#> 8                      %                swing
#> 9                      %       aucivpbextlast
#> 10                     %        aucivpbextall
#> 11                     %   aucivpbextint.last
#> 12                     %    aucivpbextint.all
#> 13                     %    aucivpbextinf.obs
#> 14                     %   aucivpbextinf.pred
#> 15                     %          aucpext.obs
#> 16                     %         aucpext.pred
#> 17                 count           count_conc
#> 18                 count    lambda.z.n.points
#> 19                    hr                start
#> 20                    hr                  end
#> 21                    hr                 tmax
#> 22                    hr                tlast
#> 23                    hr               tfirst
#> 24                    hr             mrt.last
#> 25                    hr          mrt.iv.last
#> 26                    hr                 tlag
#> 27                    hr           time_above
#> 28                    hr            half.life
#> 29                    hr  lambda.z.time.first
#> 30                    hr       thalf.eff.last
#> 31                    hr    thalf.eff.iv.last
#> 32                    hr              mrt.obs
#> 33                    hr             mrt.pred
#> 34                    hr           mrt.iv.obs
#> 35                    hr          mrt.iv.pred
#> 36                    hr           mrt.md.obs
#> 37                    hr          mrt.md.pred
#> 38                    hr        thalf.eff.obs
#> 39                    hr       thalf.eff.pred
#> 40                    hr     thalf.eff.iv.obs
#> 41                    hr    thalf.eff.iv.pred
#> 42                  1/hr             lambda.z
#> 43                  1/hr             kel.last
#> 44                  1/hr          kel.iv.last
#> 45                  1/hr              kel.obs
#> 46                  1/hr             kel.pred
#> 47                  1/hr           kel.iv.obs
#> 48                  1/hr          kel.iv.pred
#> 49                 ng/mL                   c0
#> 50                 ng/mL                 cmax
#> 51                 ng/mL                 cmin
#> 52                 ng/mL            clast.obs
#> 53                 ng/mL                  cav
#> 54                 ng/mL              ctrough
#> 55                 ng/mL               cstart
#> 56                 ng/mL                 ceoi
#> 57                 ng/mL           clast.pred
#> 58                    mg                   ae
#> 59       (ng/mL)/(mg/kg)              cmax.dn
#> 60       (ng/mL)/(mg/kg)              cmin.dn
#> 61       (ng/mL)/(mg/kg)         clast.obs.dn
#> 62       (ng/mL)/(mg/kg)        clast.pred.dn
#> 63       (ng/mL)/(mg/kg)               cav.dn
#> 64       (ng/mL)/(mg/kg)           ctrough.dn
#> 65       (mg/kg)/(ng/mL)             vss.last
#> 66       (mg/kg)/(ng/mL)          vss.iv.last
#> 67       (mg/kg)/(ng/mL)               vz.obs
#> 68       (mg/kg)/(ng/mL)              vz.pred
#> 69       (mg/kg)/(ng/mL)              vss.obs
#> 70       (mg/kg)/(ng/mL)             vss.pred
#> 71       (mg/kg)/(ng/mL)           vss.iv.obs
#> 72       (mg/kg)/(ng/mL)          vss.iv.pred
#> 73       (mg/kg)/(ng/mL)           vss.md.obs
#> 74       (mg/kg)/(ng/mL)          vss.md.pred
#> 75              hr*ng/mL              auclast
#> 76              hr*ng/mL               aucall
#> 77              hr*ng/mL          aucint.last
#> 78              hr*ng/mL     aucint.last.dose
#> 79              hr*ng/mL           aucint.all
#> 80              hr*ng/mL      aucint.all.dose
#> 81              hr*ng/mL aucabove.predose.all
#> 82              hr*ng/mL  aucabove.trough.all
#> 83              hr*ng/mL       sparse_auclast
#> 84              hr*ng/mL            aucivlast
#> 85              hr*ng/mL             aucivall
#> 86              hr*ng/mL        aucivint.last
#> 87              hr*ng/mL         aucivint.all
#> 88              hr*ng/mL           aucinf.obs
#> 89              hr*ng/mL          aucinf.pred
#> 90              hr*ng/mL       aucint.inf.obs
#> 91              hr*ng/mL  aucint.inf.obs.dose
#> 92              hr*ng/mL      aucint.inf.pred
#> 93              hr*ng/mL aucint.inf.pred.dose
#> 94              hr*ng/mL         aucivinf.obs
#> 95              hr*ng/mL        aucivinf.pred
#> 96            hr^2*ng/mL             aumclast
#> 97            hr^2*ng/mL              aumcall
#> 98            hr^2*ng/mL          aumcinf.obs
#> 99            hr^2*ng/mL         aumcinf.pred
#> 100   (hr*ng/mL)/(mg/kg)           auclast.dn
#> 101   (hr*ng/mL)/(mg/kg)            aucall.dn
#> 102   (hr*ng/mL)/(mg/kg)        aucinf.obs.dn
#> 103   (hr*ng/mL)/(mg/kg)       aucinf.pred.dn
#> 104 (hr^2*ng/mL)/(mg/kg)          aumclast.dn
#> 105 (hr^2*ng/mL)/(mg/kg)           aumcall.dn
#> 106 (hr^2*ng/mL)/(mg/kg)       aumcinf.obs.dn
#> 107 (hr^2*ng/mL)/(mg/kg)      aumcinf.pred.dn
#> 108   (mg/kg)/(hr*ng/mL)              cl.last
#> 109   (mg/kg)/(hr*ng/mL)               cl.all
#> 110   (mg/kg)/(hr*ng/mL)               cl.obs
#> 111   (mg/kg)/(hr*ng/mL)              cl.pred
#> 112        mg/(hr*ng/mL)             clr.last
#> 113        mg/(hr*ng/mL)              clr.obs
#> 114        mg/(hr*ng/mL)             clr.pred
pknca_units_table(
  concu="ng/mL", doseu="mg/kg", amountu="mg", timeu="hr",
  # Convert clearance and volume units to more understandable units with
  # automatic unit conversion
  conversions=data.frame(
    PPORRESU=c("(mg/kg)/(hr*ng/mL)", "(mg/kg)/(ng/mL)"),
    PPSTRESU=c("mL/hr/kg", "mL/kg")
  )
)
#>                 PPORRESU             PPTESTCD             PPSTRESU
#> 1               unitless            r.squared             unitless
#> 2               unitless        adj.r.squared             unitless
#> 3               fraction                    f             fraction
#> 4               fraction                  ptr             fraction
#> 5               fraction                   fe             fraction
#> 6               fraction           span.ratio             fraction
#> 7                      %             deg.fluc                    %
#> 8                      %                swing                    %
#> 9                      %       aucivpbextlast                    %
#> 10                     %        aucivpbextall                    %
#> 11                     %   aucivpbextint.last                    %
#> 12                     %    aucivpbextint.all                    %
#> 13                     %    aucivpbextinf.obs                    %
#> 14                     %   aucivpbextinf.pred                    %
#> 15                     %          aucpext.obs                    %
#> 16                     %         aucpext.pred                    %
#> 17                 count           count_conc                count
#> 18                 count    lambda.z.n.points                count
#> 19                    hr                start                   hr
#> 20                    hr                  end                   hr
#> 21                    hr                 tmax                   hr
#> 22                    hr                tlast                   hr
#> 23                    hr               tfirst                   hr
#> 24                    hr             mrt.last                   hr
#> 25                    hr          mrt.iv.last                   hr
#> 26                    hr                 tlag                   hr
#> 27                    hr           time_above                   hr
#> 28                    hr            half.life                   hr
#> 29                    hr  lambda.z.time.first                   hr
#> 30                    hr       thalf.eff.last                   hr
#> 31                    hr    thalf.eff.iv.last                   hr
#> 32                    hr              mrt.obs                   hr
#> 33                    hr             mrt.pred                   hr
#> 34                    hr           mrt.iv.obs                   hr
#> 35                    hr          mrt.iv.pred                   hr
#> 36                    hr           mrt.md.obs                   hr
#> 37                    hr          mrt.md.pred                   hr
#> 38                    hr        thalf.eff.obs                   hr
#> 39                    hr       thalf.eff.pred                   hr
#> 40                    hr     thalf.eff.iv.obs                   hr
#> 41                    hr    thalf.eff.iv.pred                   hr
#> 42                  1/hr             lambda.z                 1/hr
#> 43                  1/hr             kel.last                 1/hr
#> 44                  1/hr          kel.iv.last                 1/hr
#> 45                  1/hr              kel.obs                 1/hr
#> 46                  1/hr             kel.pred                 1/hr
#> 47                  1/hr           kel.iv.obs                 1/hr
#> 48                  1/hr          kel.iv.pred                 1/hr
#> 49                 ng/mL                   c0                ng/mL
#> 50                 ng/mL                 cmax                ng/mL
#> 51                 ng/mL                 cmin                ng/mL
#> 52                 ng/mL            clast.obs                ng/mL
#> 53                 ng/mL                  cav                ng/mL
#> 54                 ng/mL              ctrough                ng/mL
#> 55                 ng/mL               cstart                ng/mL
#> 56                 ng/mL                 ceoi                ng/mL
#> 57                 ng/mL           clast.pred                ng/mL
#> 58                    mg                   ae                   mg
#> 59       (ng/mL)/(mg/kg)              cmax.dn      (ng/mL)/(mg/kg)
#> 60       (ng/mL)/(mg/kg)              cmin.dn      (ng/mL)/(mg/kg)
#> 61       (ng/mL)/(mg/kg)         clast.obs.dn      (ng/mL)/(mg/kg)
#> 62       (ng/mL)/(mg/kg)        clast.pred.dn      (ng/mL)/(mg/kg)
#> 63       (ng/mL)/(mg/kg)               cav.dn      (ng/mL)/(mg/kg)
#> 64       (ng/mL)/(mg/kg)           ctrough.dn      (ng/mL)/(mg/kg)
#> 65       (mg/kg)/(ng/mL)             vss.last                mL/kg
#> 66       (mg/kg)/(ng/mL)          vss.iv.last                mL/kg
#> 67       (mg/kg)/(ng/mL)               vz.obs                mL/kg
#> 68       (mg/kg)/(ng/mL)              vz.pred                mL/kg
#> 69       (mg/kg)/(ng/mL)              vss.obs                mL/kg
#> 70       (mg/kg)/(ng/mL)             vss.pred                mL/kg
#> 71       (mg/kg)/(ng/mL)           vss.iv.obs                mL/kg
#> 72       (mg/kg)/(ng/mL)          vss.iv.pred                mL/kg
#> 73       (mg/kg)/(ng/mL)           vss.md.obs                mL/kg
#> 74       (mg/kg)/(ng/mL)          vss.md.pred                mL/kg
#> 75              hr*ng/mL              auclast             hr*ng/mL
#> 76              hr*ng/mL               aucall             hr*ng/mL
#> 77              hr*ng/mL          aucint.last             hr*ng/mL
#> 78              hr*ng/mL     aucint.last.dose             hr*ng/mL
#> 79              hr*ng/mL           aucint.all             hr*ng/mL
#> 80              hr*ng/mL      aucint.all.dose             hr*ng/mL
#> 81              hr*ng/mL aucabove.predose.all             hr*ng/mL
#> 82              hr*ng/mL  aucabove.trough.all             hr*ng/mL
#> 83              hr*ng/mL       sparse_auclast             hr*ng/mL
#> 84              hr*ng/mL            aucivlast             hr*ng/mL
#> 85              hr*ng/mL             aucivall             hr*ng/mL
#> 86              hr*ng/mL        aucivint.last             hr*ng/mL
#> 87              hr*ng/mL         aucivint.all             hr*ng/mL
#> 88              hr*ng/mL           aucinf.obs             hr*ng/mL
#> 89              hr*ng/mL          aucinf.pred             hr*ng/mL
#> 90              hr*ng/mL       aucint.inf.obs             hr*ng/mL
#> 91              hr*ng/mL  aucint.inf.obs.dose             hr*ng/mL
#> 92              hr*ng/mL      aucint.inf.pred             hr*ng/mL
#> 93              hr*ng/mL aucint.inf.pred.dose             hr*ng/mL
#> 94              hr*ng/mL         aucivinf.obs             hr*ng/mL
#> 95              hr*ng/mL        aucivinf.pred             hr*ng/mL
#> 96            hr^2*ng/mL             aumclast           hr^2*ng/mL
#> 97            hr^2*ng/mL              aumcall           hr^2*ng/mL
#> 98            hr^2*ng/mL          aumcinf.obs           hr^2*ng/mL
#> 99            hr^2*ng/mL         aumcinf.pred           hr^2*ng/mL
#> 100   (hr*ng/mL)/(mg/kg)           auclast.dn   (hr*ng/mL)/(mg/kg)
#> 101   (hr*ng/mL)/(mg/kg)            aucall.dn   (hr*ng/mL)/(mg/kg)
#> 102   (hr*ng/mL)/(mg/kg)        aucinf.obs.dn   (hr*ng/mL)/(mg/kg)
#> 103   (hr*ng/mL)/(mg/kg)       aucinf.pred.dn   (hr*ng/mL)/(mg/kg)
#> 104 (hr^2*ng/mL)/(mg/kg)          aumclast.dn (hr^2*ng/mL)/(mg/kg)
#> 105 (hr^2*ng/mL)/(mg/kg)           aumcall.dn (hr^2*ng/mL)/(mg/kg)
#> 106 (hr^2*ng/mL)/(mg/kg)       aumcinf.obs.dn (hr^2*ng/mL)/(mg/kg)
#> 107 (hr^2*ng/mL)/(mg/kg)      aumcinf.pred.dn (hr^2*ng/mL)/(mg/kg)
#> 108   (mg/kg)/(hr*ng/mL)              cl.last             mL/hr/kg
#> 109   (mg/kg)/(hr*ng/mL)               cl.all             mL/hr/kg
#> 110   (mg/kg)/(hr*ng/mL)               cl.obs             mL/hr/kg
#> 111   (mg/kg)/(hr*ng/mL)              cl.pred             mL/hr/kg
#> 112        mg/(hr*ng/mL)             clr.last        mg/(hr*ng/mL)
#> 113        mg/(hr*ng/mL)              clr.obs        mg/(hr*ng/mL)
#> 114        mg/(hr*ng/mL)             clr.pred        mg/(hr*ng/mL)
#>     conversion_factor
#> 1               1e+00
#> 2               1e+00
#> 3               1e+00
#> 4               1e+00
#> 5               1e+00
#> 6               1e+00
#> 7               1e+00
#> 8               1e+00
#> 9               1e+00
#> 10              1e+00
#> 11              1e+00
#> 12              1e+00
#> 13              1e+00
#> 14              1e+00
#> 15              1e+00
#> 16              1e+00
#> 17              1e+00
#> 18              1e+00
#> 19              1e+00
#> 20              1e+00
#> 21              1e+00
#> 22              1e+00
#> 23              1e+00
#> 24              1e+00
#> 25              1e+00
#> 26              1e+00
#> 27              1e+00
#> 28              1e+00
#> 29              1e+00
#> 30              1e+00
#> 31              1e+00
#> 32              1e+00
#> 33              1e+00
#> 34              1e+00
#> 35              1e+00
#> 36              1e+00
#> 37              1e+00
#> 38              1e+00
#> 39              1e+00
#> 40              1e+00
#> 41              1e+00
#> 42              1e+00
#> 43              1e+00
#> 44              1e+00
#> 45              1e+00
#> 46              1e+00
#> 47              1e+00
#> 48              1e+00
#> 49              1e+00
#> 50              1e+00
#> 51              1e+00
#> 52              1e+00
#> 53              1e+00
#> 54              1e+00
#> 55              1e+00
#> 56              1e+00
#> 57              1e+00
#> 58              1e+00
#> 59              1e+00
#> 60              1e+00
#> 61              1e+00
#> 62              1e+00
#> 63              1e+00
#> 64              1e+00
#> 65              1e+06
#> 66              1e+06
#> 67              1e+06
#> 68              1e+06
#> 69              1e+06
#> 70              1e+06
#> 71              1e+06
#> 72              1e+06
#> 73              1e+06
#> 74              1e+06
#> 75              1e+00
#> 76              1e+00
#> 77              1e+00
#> 78              1e+00
#> 79              1e+00
#> 80              1e+00
#> 81              1e+00
#> 82              1e+00
#> 83              1e+00
#> 84              1e+00
#> 85              1e+00
#> 86              1e+00
#> 87              1e+00
#> 88              1e+00
#> 89              1e+00
#> 90              1e+00
#> 91              1e+00
#> 92              1e+00
#> 93              1e+00
#> 94              1e+00
#> 95              1e+00
#> 96              1e+00
#> 97              1e+00
#> 98              1e+00
#> 99              1e+00
#> 100             1e+00
#> 101             1e+00
#> 102             1e+00
#> 103             1e+00
#> 104             1e+00
#> 105             1e+00
#> 106             1e+00
#> 107             1e+00
#> 108             1e+06
#> 109             1e+06
#> 110             1e+06
#> 111             1e+06
#> 112             1e+00
#> 113             1e+00
#> 114             1e+00
pknca_units_table(
  concu="mg/L", doseu="mg/kg", amountu="mg", timeu="hr",
  # Convert clearance and volume units to molar units (assuming
  conversions=data.frame(
    PPORRESU=c("mg/L", "(mg/kg)/(hr*ng/mL)", "(mg/kg)/(ng/mL)"),
    PPSTRESU=c("mmol/L", "mL/hr/kg", "mL/kg"),
    # Manual conversion of concentration units from ng/mL to mmol/L (assuming
    # a molecular weight of 138.121 g/mol)
    conversion_factor=c(1/138.121, NA, NA)
  )
)
#> Warning: The following unit conversions were supplied but do not match any units to convert: '(mg/kg)/(hr*ng/mL)', '(mg/kg)/(ng/mL)'
#>                PPORRESU             PPTESTCD            PPSTRESU
#> 1              unitless            r.squared            unitless
#> 2              unitless        adj.r.squared            unitless
#> 3              fraction                    f            fraction
#> 4              fraction                  ptr            fraction
#> 5              fraction                   fe            fraction
#> 6              fraction           span.ratio            fraction
#> 7                     %             deg.fluc                   %
#> 8                     %                swing                   %
#> 9                     %       aucivpbextlast                   %
#> 10                    %        aucivpbextall                   %
#> 11                    %   aucivpbextint.last                   %
#> 12                    %    aucivpbextint.all                   %
#> 13                    %    aucivpbextinf.obs                   %
#> 14                    %   aucivpbextinf.pred                   %
#> 15                    %          aucpext.obs                   %
#> 16                    %         aucpext.pred                   %
#> 17                count           count_conc               count
#> 18                count    lambda.z.n.points               count
#> 19                   hr                start                  hr
#> 20                   hr                  end                  hr
#> 21                   hr                 tmax                  hr
#> 22                   hr                tlast                  hr
#> 23                   hr               tfirst                  hr
#> 24                   hr             mrt.last                  hr
#> 25                   hr          mrt.iv.last                  hr
#> 26                   hr                 tlag                  hr
#> 27                   hr           time_above                  hr
#> 28                   hr            half.life                  hr
#> 29                   hr  lambda.z.time.first                  hr
#> 30                   hr       thalf.eff.last                  hr
#> 31                   hr    thalf.eff.iv.last                  hr
#> 32                   hr              mrt.obs                  hr
#> 33                   hr             mrt.pred                  hr
#> 34                   hr           mrt.iv.obs                  hr
#> 35                   hr          mrt.iv.pred                  hr
#> 36                   hr           mrt.md.obs                  hr
#> 37                   hr          mrt.md.pred                  hr
#> 38                   hr        thalf.eff.obs                  hr
#> 39                   hr       thalf.eff.pred                  hr
#> 40                   hr     thalf.eff.iv.obs                  hr
#> 41                   hr    thalf.eff.iv.pred                  hr
#> 42                 1/hr             lambda.z                1/hr
#> 43                 1/hr             kel.last                1/hr
#> 44                 1/hr          kel.iv.last                1/hr
#> 45                 1/hr              kel.obs                1/hr
#> 46                 1/hr             kel.pred                1/hr
#> 47                 1/hr           kel.iv.obs                1/hr
#> 48                 1/hr          kel.iv.pred                1/hr
#> 49                 mg/L                   c0              mmol/L
#> 50                 mg/L                 cmax              mmol/L
#> 51                 mg/L                 cmin              mmol/L
#> 52                 mg/L            clast.obs              mmol/L
#> 53                 mg/L                  cav              mmol/L
#> 54                 mg/L              ctrough              mmol/L
#> 55                 mg/L               cstart              mmol/L
#> 56                 mg/L                 ceoi              mmol/L
#> 57                 mg/L           clast.pred              mmol/L
#> 58                   mg                   ae                  mg
#> 59       (mg/L)/(mg/kg)              cmax.dn      (mg/L)/(mg/kg)
#> 60       (mg/L)/(mg/kg)              cmin.dn      (mg/L)/(mg/kg)
#> 61       (mg/L)/(mg/kg)         clast.obs.dn      (mg/L)/(mg/kg)
#> 62       (mg/L)/(mg/kg)        clast.pred.dn      (mg/L)/(mg/kg)
#> 63       (mg/L)/(mg/kg)               cav.dn      (mg/L)/(mg/kg)
#> 64       (mg/L)/(mg/kg)           ctrough.dn      (mg/L)/(mg/kg)
#> 65       (mg/kg)/(mg/L)             vss.last      (mg/kg)/(mg/L)
#> 66       (mg/kg)/(mg/L)          vss.iv.last      (mg/kg)/(mg/L)
#> 67       (mg/kg)/(mg/L)               vz.obs      (mg/kg)/(mg/L)
#> 68       (mg/kg)/(mg/L)              vz.pred      (mg/kg)/(mg/L)
#> 69       (mg/kg)/(mg/L)              vss.obs      (mg/kg)/(mg/L)
#> 70       (mg/kg)/(mg/L)             vss.pred      (mg/kg)/(mg/L)
#> 71       (mg/kg)/(mg/L)           vss.iv.obs      (mg/kg)/(mg/L)
#> 72       (mg/kg)/(mg/L)          vss.iv.pred      (mg/kg)/(mg/L)
#> 73       (mg/kg)/(mg/L)           vss.md.obs      (mg/kg)/(mg/L)
#> 74       (mg/kg)/(mg/L)          vss.md.pred      (mg/kg)/(mg/L)
#> 75              hr*mg/L              auclast             hr*mg/L
#> 76              hr*mg/L               aucall             hr*mg/L
#> 77              hr*mg/L          aucint.last             hr*mg/L
#> 78              hr*mg/L     aucint.last.dose             hr*mg/L
#> 79              hr*mg/L           aucint.all             hr*mg/L
#> 80              hr*mg/L      aucint.all.dose             hr*mg/L
#> 81              hr*mg/L aucabove.predose.all             hr*mg/L
#> 82              hr*mg/L  aucabove.trough.all             hr*mg/L
#> 83              hr*mg/L       sparse_auclast             hr*mg/L
#> 84              hr*mg/L            aucivlast             hr*mg/L
#> 85              hr*mg/L             aucivall             hr*mg/L
#> 86              hr*mg/L        aucivint.last             hr*mg/L
#> 87              hr*mg/L         aucivint.all             hr*mg/L
#> 88              hr*mg/L           aucinf.obs             hr*mg/L
#> 89              hr*mg/L          aucinf.pred             hr*mg/L
#> 90              hr*mg/L       aucint.inf.obs             hr*mg/L
#> 91              hr*mg/L  aucint.inf.obs.dose             hr*mg/L
#> 92              hr*mg/L      aucint.inf.pred             hr*mg/L
#> 93              hr*mg/L aucint.inf.pred.dose             hr*mg/L
#> 94              hr*mg/L         aucivinf.obs             hr*mg/L
#> 95              hr*mg/L        aucivinf.pred             hr*mg/L
#> 96            hr^2*mg/L             aumclast           hr^2*mg/L
#> 97            hr^2*mg/L              aumcall           hr^2*mg/L
#> 98            hr^2*mg/L          aumcinf.obs           hr^2*mg/L
#> 99            hr^2*mg/L         aumcinf.pred           hr^2*mg/L
#> 100   (hr*mg/L)/(mg/kg)           auclast.dn   (hr*mg/L)/(mg/kg)
#> 101   (hr*mg/L)/(mg/kg)            aucall.dn   (hr*mg/L)/(mg/kg)
#> 102   (hr*mg/L)/(mg/kg)        aucinf.obs.dn   (hr*mg/L)/(mg/kg)
#> 103   (hr*mg/L)/(mg/kg)       aucinf.pred.dn   (hr*mg/L)/(mg/kg)
#> 104 (hr^2*mg/L)/(mg/kg)          aumclast.dn (hr^2*mg/L)/(mg/kg)
#> 105 (hr^2*mg/L)/(mg/kg)           aumcall.dn (hr^2*mg/L)/(mg/kg)
#> 106 (hr^2*mg/L)/(mg/kg)       aumcinf.obs.dn (hr^2*mg/L)/(mg/kg)
#> 107 (hr^2*mg/L)/(mg/kg)      aumcinf.pred.dn (hr^2*mg/L)/(mg/kg)
#> 108   (mg/kg)/(hr*mg/L)              cl.last   (mg/kg)/(hr*mg/L)
#> 109   (mg/kg)/(hr*mg/L)               cl.all   (mg/kg)/(hr*mg/L)
#> 110   (mg/kg)/(hr*mg/L)               cl.obs   (mg/kg)/(hr*mg/L)
#> 111   (mg/kg)/(hr*mg/L)              cl.pred   (mg/kg)/(hr*mg/L)
#> 112        mg/(hr*mg/L)             clr.last        mg/(hr*mg/L)
#> 113        mg/(hr*mg/L)              clr.obs        mg/(hr*mg/L)
#> 114        mg/(hr*mg/L)             clr.pred        mg/(hr*mg/L)
#>     conversion_factor
#> 1         1.000000000
#> 2         1.000000000
#> 3         1.000000000
#> 4         1.000000000
#> 5         1.000000000
#> 6         1.000000000
#> 7         1.000000000
#> 8         1.000000000
#> 9         1.000000000
#> 10        1.000000000
#> 11        1.000000000
#> 12        1.000000000
#> 13        1.000000000
#> 14        1.000000000
#> 15        1.000000000
#> 16        1.000000000
#> 17        1.000000000
#> 18        1.000000000
#> 19        1.000000000
#> 20        1.000000000
#> 21        1.000000000
#> 22        1.000000000
#> 23        1.000000000
#> 24        1.000000000
#> 25        1.000000000
#> 26        1.000000000
#> 27        1.000000000
#> 28        1.000000000
#> 29        1.000000000
#> 30        1.000000000
#> 31        1.000000000
#> 32        1.000000000
#> 33        1.000000000
#> 34        1.000000000
#> 35        1.000000000
#> 36        1.000000000
#> 37        1.000000000
#> 38        1.000000000
#> 39        1.000000000
#> 40        1.000000000
#> 41        1.000000000
#> 42        1.000000000
#> 43        1.000000000
#> 44        1.000000000
#> 45        1.000000000
#> 46        1.000000000
#> 47        1.000000000
#> 48        1.000000000
#> 49        0.007240029
#> 50        0.007240029
#> 51        0.007240029
#> 52        0.007240029
#> 53        0.007240029
#> 54        0.007240029
#> 55        0.007240029
#> 56        0.007240029
#> 57        0.007240029
#> 58        1.000000000
#> 59        1.000000000
#> 60        1.000000000
#> 61        1.000000000
#> 62        1.000000000
#> 63        1.000000000
#> 64        1.000000000
#> 65        1.000000000
#> 66        1.000000000
#> 67        1.000000000
#> 68        1.000000000
#> 69        1.000000000
#> 70        1.000000000
#> 71        1.000000000
#> 72        1.000000000
#> 73        1.000000000
#> 74        1.000000000
#> 75        1.000000000
#> 76        1.000000000
#> 77        1.000000000
#> 78        1.000000000
#> 79        1.000000000
#> 80        1.000000000
#> 81        1.000000000
#> 82        1.000000000
#> 83        1.000000000
#> 84        1.000000000
#> 85        1.000000000
#> 86        1.000000000
#> 87        1.000000000
#> 88        1.000000000
#> 89        1.000000000
#> 90        1.000000000
#> 91        1.000000000
#> 92        1.000000000
#> 93        1.000000000
#> 94        1.000000000
#> 95        1.000000000
#> 96        1.000000000
#> 97        1.000000000
#> 98        1.000000000
#> 99        1.000000000
#> 100       1.000000000
#> 101       1.000000000
#> 102       1.000000000
#> 103       1.000000000
#> 104       1.000000000
#> 105       1.000000000
#> 106       1.000000000
#> 107       1.000000000
#> 108       1.000000000
#> 109       1.000000000
#> 110       1.000000000
#> 111       1.000000000
#> 112       1.000000000
#> 113       1.000000000
#> 114       1.000000000